% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_score_data.R
\name{add_cci_data}
\alias{add_cci_data}
\title{Adds the CCI score endpoint data to the score data}
\usage{
add_cci_data(score_data, score_type, pheno_data, icd_data, study_setup)
}
\arguments{
\item{score_data}{A data.frame. The current score data for the study setup.}

\item{score_type}{A string (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index,
\item \code{EI} for Elixhauser index,
\item \code{PRS} for polygenic risk score,
\item \code{PheRS} for PheRS results score,
\item \code{MED} for medication data, and
\item \code{EDU} for educational level. The column for this needs to
be called \code{EDU} in the \code{pheno_data}.
\item \code{ZIP} for zip scores,
}}

\item{pheno_data}{A data.frame. The phenotypic data on the individuals.
Should at least contain the columns as described  in
\link{get_relevant_pheno_data_cols}.}

\item{icd_data}{A data.frame. The ICD codes of the individuals. Needs at least the
columns as described in \link{get_study_cci_data}.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup]{Istudy::study_setup}.}
}
\value{
A tibble. The score data with added columns \code{CCI}, or \code{EI} depending
on the \code{score_type} used.
}
\description{
Adds the CCI score endpoint data to the score data
}
\author{
Kira E. Detrois
}
