% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{add_cidx_res_row}
\alias{add_cidx_res_row}
\title{Adds a row to the C-index results tibble}
\usage{
add_cidx_res_row(c_idxs_tib, c_idx_res, endpt)
}
\arguments{
\item{endpt}{A character. The column name of the current endpoint of
interest.}

\item{endpt_hrs_tib}{A tibble with the results for previous endpts.}

\item{coxph_mdl}{A Cox-PH model.}

\item{elig_indv}{A tibble. The individuals which were eligble under
the current study setup. Needs to at least
contain the column defined in \code{endpt}.}
}
\value{
A tibble. The updated \code{endpt_hrs_tib} with the added results for
the current endpoint.
}
\description{
Adds a row to the C-index results tibble
}
\author{
Kira E. Detrois
}
