% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extract_res.R
\name{add_coxph_row}
\alias{add_coxph_row}
\title{Adds a row to the results tibble}
\usage{
add_coxph_row(all_coxph_res, coxph_res_tib, score_type, endpt, elig_indv)
}
\arguments{
\item{all_coxph_res}{A tibble. The results for the previous endpoints.}

\item{coxph_res_tib}{A tibble with the results for the current endpoints.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{endpt}{A character. The current enpoint of interest.}

\item{elig_indv}{A tibble. The individuals which were eligble under
the current study setup. Needs to at least contain
the column defined in \code{endpt}.}
}
\value{
A tibble. TODO
}
\description{
Adds a row to the results tibble
}
\author{
Kira E. Detrois
}
