% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_score_data.R
\name{add_edu_data}
\alias{add_edu_data}
\title{Adds the education data to the score data}
\usage{
add_edu_data(score_data, score_type, pheno_data)
}
\arguments{
\item{score_data}{A data.frame. The current score data for the study setup.}

\item{score_type}{A string (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index,
\item \code{EI} for Elixhauser index,
\item \code{PRS} for polygenic risk score,
\item \code{PheRS} for PheRS results score,
\item \code{MED} for medication data, and
\item \code{EDU} for educational level. The column for this needs to
be called \code{ISCED_2011} in the \code{pheno_data}.
\item \code{ZIP} for zip scores,
}}

\item{pheno_data}{A data.frame. The phenotypic data on the individuals.
Should at least contain the columns as described  in
\link{get_relevant_pheno_data_cols}.}
}
\value{
A tibble. The score data with added column \code{EDU}.
}
\description{
The education data comes from the phenotype data and gets
mapped from ISCED 2011 to the FinnGen R10 age modes, see
function \link{get_study_edu_data}.
}
\author{
Kira E. Detrois
}
