% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_score_data.R
\name{add_med_endpt_data}
\alias{add_med_endpt_data}
\title{Adds the medication score endpoint data to the score data}
\usage{
add_med_endpt_data(score_data, score_type, pheno_data, study_setup, atc_data)
}
\arguments{
\item{score_data}{A data.frame. The current score data for the study setup.}

\item{score_type}{A string (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index,
\item \code{EI} for Elixhauser index,
\item \code{PRS} for polygenic risk score,
\item \code{PheRS} for PheRS results score,
\item \code{MED} for medication data, and
\item \code{EDU} for educational level. The column for this needs to
be called \code{EDU} in the \code{pheno_data}.
\item \code{ZIP} for zip scores,
}}

\item{pheno_data}{A data.frame. The phenotypic data on the individuals.
Should at least contain the columns as described  in
\link{get_relevant_pheno_data_cols}.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup]{Istudy::study_setup}.}

\item{atc_data}{A data.frame. The ATC codes of the individuals. Needs at least the
columns as described in \link{get_study_med_data}.}
}
\value{
A tibble. The score data with added column \code{MED}.
}
\description{
Adds the medication score endpoint data to the score data
}
\author{
Kira E. Detrois
}
