% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_score_data.R
\name{add_prs_endpt_data}
\alias{add_prs_endpt_data}
\title{Adds the PRS endpoint data to the score data}
\usage{
add_prs_endpt_data(score_data, score_type, prs_data, endpt)
}
\arguments{
\item{score_data}{A data.frame. The current score data for the study setup.}

\item{score_type}{a character (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index,
\item \code{EI} for Elixhauser index,
\item \code{PRS} for polygenic risk score,
\item \code{PheRS} for PheRS results score,
\item \code{MED} for medication data, and
\item \code{EDU} for educational level. The column for this needs to
be called \code{ISCED_2011} in the \code{pheno_data}.
\item \code{ZIP} for zip scores,
}}

\item{prs_data}{A data.frame. The PRS scores of the individuals. Needs at least the
columns as described in \link{get_prs_endpt_data}.}

\item{endpt}{A character. The current endpoint.
Required if \code{PRS} or \code{PheRS} is present in \code{score_type}.}
}
\value{
A tibble. The score data with added column \code{PRS}.
}
\description{
Adds the PRS endpoint data to the score data
}
\author{
Kira E. Detrois
}
