% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_back_study_hrs.R
\name{calc_endpt_back_study_hrs}
\alias{calc_endpt_back_study_hrs}
\title{Calcualtes HR from a Cox-PH model for each exposure age and endpoint}
\usage{
calc_endpt_back_study_hrs(
  elig_score_data,
  score_type,
  back_study,
  obs_end,
  endpt_studies,
  downsample_fctr = NULL,
  ancs = NA_character_,
  covs = c("SEX", "YEAR_OF_BIRTH"),
  bin_cut = 1,
  min_indvs = 5,
  write_res = FALSE,
  res_dir = NULL
)
}
\arguments{
\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{endpt_studies}{A vector of S4 classes representing the study setups.}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type == CCI}.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}

\item{score_ages_data}{A named vector of tibbles.
The score data for each indiviuals for different
exposure periods.}

\item{endpts}{A string. The column names of the endpoints of
interest.}

\item{exp_ages}{An integer (vector). Ages at which exposure period
starts (in years).}

\item{exp_len}{An integer. Length of the exposure period ^
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}
}
\description{
For more details see \link{calc_endpt_studies_hrs} function documentation.
}
\author{
Kira E. Detrois
}
