% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_study_hr.R
\name{calc_endpt_study_hr}
\alias{calc_endpt_study_hr}
\title{Calcualtes HR from a Cox-PH model for a survival analysis setup}
\usage{
calc_endpt_study_hr(surv_ana)
}
\arguments{
\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\value{
A Cox-PH model.
}
\description{
Calcualtes HR from a Cox-PH model for a survival analysis setup
based on the risk groups.
}
\details{
For i.e. a study with endpoint \code{J10_ASTHMA} and
\code{covs = c("SEX", "YEAR_OF_BIRTH")} the model would be
\code{Surv(J10_ASTHMA_AGE_DAYS, J10_ASTHMA) ~ SCORE_GROUP + SEX + YEAR_OF_BIRTH}.
}
\author{
Kira E. Detrois
}
