% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_data.R
\name{calc_exp_end_age}
\alias{calc_exp_end_age}
\title{Calcualtes the age of individuals at exposure end}
\usage{
calc_exp_end_age(elig_indv)
}
\arguments{
\item{elig_indv}{A dataframe of individuals with columns
\code{ID}, \code{DATE_OF_BIRTH}, and \code{EXP_END_DATE}.}
}
\value{
tibble with columns \code{ID} and \code{EXP_END},
which is the ID of the individual and the corresponding
calculated exposure end date.
}
\description{
Calculates the length of the exposure period in years
from birth until the end date of the exposure period.
}
\author{
Kira E. Detrois
}
