% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_data.R
\name{calc_exp_start_age}
\alias{calc_exp_start_age}
\title{Calcualtes the age of individuals at exposure start}
\usage{
calc_exp_start_age(elig_indv)
}
\arguments{
\item{elig_indv}{A data.frame of individuals with columns
\code{ID}, \code{DATE_OF_BIRTH}, and \code{EXP_START_DATE}.}
}
\value{
tibble with columns \code{ID} and \code{EXP_START},
which is the ID of the individual and the corresponding
calculated exposure start date.
}
\description{
Calculates the length of the exposure period in (exact) years
from birth until the start date of the exposure period.
}
\author{
Kira E. Detrois
}
