% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_studies_hrs.R
\name{calc_studies_hrs}
\alias{calc_studies_hrs}
\title{Calcualtes HR from a Cox-PH model for each endpoint}
\usage{
calc_studies_hrs(
  pheno_data,
  score_data,
  score_type,
  studies,
  covs = c("SEX", "YEAR_OF_BIRTH"),
  bin_cut = 1,
  write_res = FALSE,
  res_dir = NA
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, the columns specified in \code{covs} and
i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE} where the
columns are the study endpoint and date, which will
differ depending on the input variable \code{endpts}.}

\item{score_data}{A data.frame. Needs at least column \code{SCORE} if
\code{score_type = CCI}, or i.e. \code{J10_ASTHMA_PRS} for
\code{score_type = PRS} depending on the chosen \code{endpts}.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{studies}{A vector of S4 classes representing the study setups.}

\item{covs}{A character. The column names of the covariates to add to
the predictor of the Cox-PH model.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type} is \code{CCI}.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\description{
Calcualtes HR from a Cox-PH model for each endpoint
}
\author{
Kira E. Detrois
}
