% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{check_and_get_file_path}
\alias{check_and_get_file_path}
\title{Creats the file directory and name for the different result types}
\usage{
check_and_get_file_path(ana_details, res_type)
}
\arguments{
\item{res_type}{A character. The results type.}
}
\value{
A character. The file name.
}
\description{
Creats the file directory and name for the different result types
}
\author{
Kira E. Detrois
}
