% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{check_file_exists}
\alias{check_file_exists}
\title{Checks whether the log file already exists}
\usage{
check_file_exists(file_path)
}
\value{
A boolean. Whether the file exists.
}
\description{
Checks whether the log file already exists
}
\author{
Kira E. Detrois
}
