% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{check_has_mid_group}
\alias{check_has_mid_group}
\title{Checks whether there is a 40-60\% group in the score groups}
\usage{
check_has_mid_group(score_group_tbl)
}
\arguments{
\item{score_group_tbl}{A named character.
The score cutoffs that should be written to the log.}
}
\value{
A bollean. Indicates whether the group exists.
}
\description{
Checks whether there is a 40-60\% group in the score groups
}
\author{
Kira E. Detrois
}
