% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_logs.R
\name{check_res_dir}
\alias{check_res_dir}
\title{Checks whether to write a results file and if the directory exists}
\usage{
check_res_dir(write_res = NULL, res_dir = NULL)
}
\arguments{
\item{write_res}{A boolean. Whether to write log files.}

\item{res_dir}{A character. The path to the results directory.}
}
\value{
A boolean. Whether to write the results to the \code{res_dir}.
}
\description{
If the directory does not exists tries to create it, recursively.
}
\author{
Kira E. Detrois
}
