% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv_studies.R
\name{create_crnt_study_from_file}
\alias{create_crnt_study_from_file}
\title{Creates a study object}
\usage{
create_crnt_study_from_file(surv_ana, study_setup, endpt)
}
\arguments{
\item{surv_ana}{An S4 \code{surv_ana} object. The current survival analysis setup.
See class definition \link{surv_ana}.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}

\item{endpt}{A character. The current endpoint.}
}
\value{
An S4 study object with the current endpoint.
}
\description{
Reads in previously created phenotype-score files. If the file doesn't exist
for the endpoint creates an empty tibble. Then creates a \code{study} object for
each selected endpoint and given \code{study_setup}.
}
\author{
Kira E. Detrois
}
