% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{create_empty_cidx_res}
\alias{create_empty_cidx_res}
\title{Creates an empty tibble for the c-index results}
\usage{
create_empty_cidx_res()
}
\value{
An empty tibble with the following columns:
\itemize{
\item ENDPOINT: endpoint name
\item EXP_AGE: exposure age
\item SURV_MODEL: A string describing the the survival model
\item N_CASES: a numeric column for the number of cases
\item N_CONTROLS: a numeric column for the number of controls
\item C_IDX: a numeric column for the c-index
\item SE: a numeric column for the standard error of the c-index
\item C_IDX_CI_NEG: a numeric column for the lower bound of the
c-index's confidence interval
\item C_IDX_CI_POS: a numeric column for the upper bound of the
c-index's confidence interval
}
}
\description{
Creates an empty tibble for the c-index results
}
\author{
Kira E. Detrois
}
