% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_extract_res.R
\name{create_empty_coxph_res_tib}
\alias{create_empty_coxph_res_tib}
\title{Creates an empty tibble for the results}
\usage{
create_empty_coxph_res_tib()
}
\value{
An empty tibble with all relevant columns for the final results.
}
\description{
Creates an empty tibble for the results
}
\author{
Kira E. Detrois
}
