% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{create_empty_endpt_hrs_tib}
\alias{create_empty_endpt_hrs_tib}
\title{Creates an empty tibble for the Cox-PH HR results}
\usage{
create_empty_endpt_hrs_tib()
}
\value{
An empty tibble with all relevant columns for the final
results.
}
\description{
Creates an empty tibble for the Cox-PH HR results
}
\author{
Kira E. Detrois
}
