% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endpts_study_objs.R
\name{create_endpt_study_obj}
\alias{create_endpt_study_obj}
\title{Creates an S4 study object}
\usage{
create_endpt_study_obj(
  pheno_data,
  study_type = "forward",
  endpt = "J10_ASTHMA",
  exp_age = NULL,
  exp_len = 10,
  wash_len = 2,
  obs_len = 8,
  obs_end = NULL,
  downsample_fctr = NA_integer_,
  ancs = c("EAS")
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, the columns specified in \code{covs} and
i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_AGE_DAYS} where
the columns are the study endpoint and date, which
will differ depending on the chosen \code{endpts}.
The phenotypic data on the individuals.}

\item{study_type}{A character. Can be either \code{forward} or \code{backward}.
\code{forward} considers individuals of a certain age
and a set exposure, washout, and observation periods
calcualted onwards from this age. It can simply be created
by creating a S4 study object.
\code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point. The exposure period will
be different for each individual depending on their birth
date. This setup can be created, using the function
\code{\link{get_backward_study}}.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period ^
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}

\item{obs_end}{A Date. The end of the observation period. Needed for
\code{backward} studies. If not set the end will be
the most recent date across the diagnosis columns.}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{ancs}{A character (vector). The ancestries to consider.}
}
\value{
An S4 study object.
}
\description{
Creates an S4 study object
}
\author{
Kira E. Detrois
}
