% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_age_exp_studies.R
\name{create_endpts_study_objs}
\alias{create_endpts_study_objs}
\title{Creates a vector of study objects for the differen endpoints}
\usage{
create_endpts_study_objs(
  endpts,
  exp_age = 30,
  exp_len = 10,
  wash_len = 2,
  obs_len = 8,
  downsample_fctr = NA_real_
)
}
\arguments{
\item{endpts}{A string. The column names of the endpoints of
interest.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}
}
\description{
Creates a vector of study objects for the differen endpoints
}
\author{
Kira E. Detrois
}
