% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_log.R
\name{create_log_file}
\alias{create_log_file}
\title{Creates log file}
\usage{
create_log_file(.Object)
}
\arguments{
\item{.Object}{An S4 \code{surv_ana} object. The current survival analysis setup.
See class definition \link{surv_ana}.}
}
\value{
A string. The file path for the error file.
}
\description{
Creates log file
}
\author{
Kira E. Detrois
}
