% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv_studies.R
\name{create_pheno_score_files}
\alias{create_pheno_score_files}
\title{Create phenotype score files}
\usage{
create_pheno_score_files(
  study_setup,
  endpts,
  surv_ana,
  score_type,
  pheno_data,
  endpts_indvs_mat,
  icd_data,
  atc_data,
  prs_data,
  phers_data,
  zip_data
)
}
\arguments{
\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup]{Istudy::study_setup}.}

\item{endpts}{A string (vector). The endpoints of interest.}

\item{surv_ana}{An S4 \code{surv_ana} object. The current survival analysis setup.
See class definition \link{surv_ana}.}

\item{score_type}{A string (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index,
\item \code{EI} for Elixhauser index,
\item \code{PRS} for polygenic risk score,
\item \code{PheRS} for PheRS results score,
\item \code{MED} for medication data, and
\item \code{EDU} for educational level. The column for this needs to
be called \code{EDU} in the \code{pheno_data}.
\item \code{ZIP} for zip scores,
}}

\item{pheno_data}{A data.frame. The phenotypic data on the individuals.
Should at least contain the columns as described  in
\link{get_relevant_pheno_data_cols}.}

\item{endpts_indvs_mat}{A data.frame. A matrix with the individuals that can
be used for each endpoint. Contains a column of individual
IDs and a binary column for each endpoint.}

\item{icd_data}{A data.frame. The ICD codes of the individuals. Needs at least the
columns as described in \link{get_study_cci_data}.}

\item{atc_data}{A data.frame. The ATC codes of the individuals. Needs at least the
columns as described in \link{get_study_med_data}.}

\item{prs_data}{A data.frame. The PRS scores of the individuals. Needs at least the
columns as described in \link{get_prs_endpt_data}.}

\item{phers_data}{A data.frame. The PheRS scores of the individuals. Needs at least
the columns as described in \link{get_phers_endpt_data}.}

\item{zip_data}{A data.frame. The ZIP scores of the individuals. Needs at least the
columns as described in \link{get_zip_data}.}
}
\description{
Reads in the phenotype data and relevant score data, defined by the selected score
types and writes the phenotype-score data to files
}
\details{
The \code{endpts_indvs_mat} is mostly relevant for FinnGen where we have some overalp for
the PRS studies and training and testing sets. However, it could
also be relevant for the training and testing of the PheRS. See function
\link{get_crnt_endpt_data}.
}
\author{
Kira E. Detrois
}
