% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_preds_vec.R
\name{create_preds_vec}
\alias{create_preds_vec}
\title{Creates a vector of the column names of all predictors of the model}
\usage{
create_preds_vec(score_type, pred_score, covs)
}
\description{
Creates a vector of the column names of all predictors of the model
}
