% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{extract_coxph_res}
\alias{extract_coxph_res}
\title{Extracts the relevant results from the cox-ph model}
\usage{
extract_coxph_res(coxph_res)
}
\arguments{
\item{coxph_res}{A coxph model.}
}
\value{
A list(\code{beta}, \code{std_err}, \code{p_val}, \code{HR}, \code{CI}, \code{groups}).
}
\description{
Extracts the relevant results from the cox-ph model
}
