% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{filter_out_missing_hrs}
\alias{filter_out_missing_hrs}
\title{Filters out HRs that are NA, or where either bound of the CI is infinite}
\usage{
filter_out_missing_hrs(coxph_hrs)
}
\description{
Filters out HRs that are NA, or where either bound of the CI is infinite
}
