% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_obj.R
\name{get_age_sd_hr_ggplot}
\alias{get_age_sd_hr_ggplot}
\title{Creats ggplot for forwar study setup}
\usage{
get_age_sd_hr_ggplot(coxph_hrs, study_setup, preds, endpt, min_y, max_y)
}
\arguments{
\item{coxph_hrs}{A dataframe. The results of the Cox-PH model for
the current endpoint.}

\item{min_y}{A numeric. Minimum value for the y-axis.}

\item{max_y}{A numeric. Maximum value for the y-axis.}

\item{ana_details}{A list. Details of the current analysis such as
study type, predictors, etc.}
}
\value{
A ggplot object.
}
\description{
Creates the ggplot for the HRs vs. observation age for an endpoint
in a forward study.
}
\author{
Kira E. Detrois
}
