% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_all_preds_sorted}
\alias{get_all_preds_sorted}
\title{Get All Preds Sorted}
\usage{
get_all_preds_sorted(score_type, covs)
}
\arguments{
\item{score_type}{A character. The score types used in the Cox-PH model.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\value{
A character vector of all predictors being used in the survival analysis,
sorted by predictor name.
}
\description{
Returns a sorted character vector of all predictors being used in the survival
analysis. Interaction predictors are sorted separately from non-interaction
predictors.
}
\author{
Kira E. Detrois
}
