% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_files.R
\name{get_ana_details_from_surv_ana}
\alias{get_ana_details_from_surv_ana}
\title{Creates a list if the details from the survival analysis}
\usage{
get_ana_details_from_surv_ana(surv_ana)
}
\description{
For plotting and file saving purposes, so that one can easily use
the same functions also without creating a surv_ana object.
}
