% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_and_filter_endpt_scores.R
\name{get_and_filter_endpt_scores}
\alias{get_and_filter_endpt_scores}
\title{Renames the score column to \code{SCORE} and filters out NAs}
\usage{
get_and_filter_endpt_scores(score_data, score_type, endpt)
}
\arguments{
\item{score_data}{A data.frame with at least columns \code{SCORE}, or
i.e. \code{J10_ASTHMA_PRS} for endpoint \code{J10_ASTHMA}
and score type \code{PRS}.}

\item{score_type}{A character. The score type. At the moment the two
options are \code{CCI}, and \code{PRS}.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}
}
\value{
A tibble. The renamed and filtered score data.
}
\description{
For the PRS data the score columsn have names in the
form of \code{J10_ASTHMA_PRS}. The function renames the
current PRS column of interest to \code{SCORE}. Then it
filters out all NAs in the column.
}
