% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_caption}
\alias{get_caption}
\title{Returns the plot caption for a HRs plot}
\usage{
get_caption(study_setup, preds)
}
\arguments{
\item{ana_details}{A list. Details of the survival analysis including
\code{study_type}, \code{obs_len}, \code{wash_len}, \code{exp_len}, \code{preds}.}
}
\value{
A string. The plot caption.
}
\description{
The function generates a caption that describes the study setup
and the predictor variables used in the Cox-PH model.
}
\author{
Kira E. Detrois
}
