% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coxph_res.R
\name{get_CI}
\alias{get_CI}
\title{Calculate 95\% Confidence Interval}
\usage{
get_CI(ML, SE)
}
\arguments{
\item{ML}{A numeric vector of estimates.}

\item{SE}{A numeric vector of standard errors.}
}
\value{
A list with two elements, \code{neg} (negative limits of the confidence interval) and
\code{pos} (positive limits of the confidence interval).
}
\description{
This function calculates the 95\% confidence interval for a set of
estimates using the standard error of the estimates.
}
\author{
Kira E. Detrois
}
