% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_cidx_res_file_name}
\alias{get_cidx_res_file_name}
\title{Creats the file name for Cox-PH model results file}
\usage{
get_cidx_res_file_name(surv_ana)
}
\value{
A character. The file name.
}
\description{
Creats the file name for Cox-PH model results file
}
\author{
Kira E. Detrois
}
