% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cors.R
\name{get_cor_data}
\alias{get_cor_data}
\title{Very inelegant way to get the pearsons' correlation between variables depending on modles run}
\usage{
get_cor_data(pheno_score_data, score_type, endpt, cor_data)
}
\description{
Very inelegant way to get the pearsons' correlation between variables depending on modles run
}
