% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{get_coxph_formula}
\alias{get_coxph_formula}
\title{Creates the forumla string for the Cox-PH model.}
\usage{
get_coxph_formula(endpt, covs, pred_score = "SCORE_GROUP")
}
\arguments{
\item{endpt}{A character. The column name of the current endpoint of
interest.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\value{
A character. The formula string for the Cox-PH model.
}
\description{
Creates the forumla string for the Cox-PH model.
}
