% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coxph_mdl.R
\name{get_coxph_mdl}
\alias{get_coxph_mdl}
\title{Fits the Cox-PH model for a given endpoint and covariates}
\usage{
get_coxph_mdl(pheno_score_data, endpt, covs, pred_score = "SCORE_GROUP")
}
\arguments{
\item{pheno_score_data}{A data.frame with at least the columns:
\code{SCORE_GROUP}, the columns specified in \code{covs}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_AGE_DAYS}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\value{
A Cox-PH model object or \code{NULL} if the model couldn't be fit.
}
\description{
For i.e. \code{endpt = "J10_ASTHMA"} and \code{covs = c("SEX", "YEAR_OF_BIRTH")}
the model would be
\code{Surv(J10_ASTHMA_AGE_DAYS, J10_ASTHMA) ~ SCORE_GROUP + SEX + YEAR_OF_BIRTH}.
}
\author{
Kira E. Detrois
}
