% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_coxph_res_file_name}
\alias{get_coxph_res_file_name}
\title{Creats the file name for Cox-PH model results file}
\usage{
get_coxph_res_file_name(study, score_type, bin_cut, obs_end = NULL)
}
\arguments{
\item{study}{An S4 class representing the study setup.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type == CCI}.}
}
\value{
A character. The file name.
}
\description{
Creats the file name for Cox-PH model results file
}
\author{
Kira E. Detrois
}
