% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_save.R
\name{get_endpt_fig_height}
\alias{get_endpt_fig_height}
\title{Get the height of the plot depending on number of predictor variables}
\usage{
get_endpt_fig_height(coxph_vars)
}
\arguments{
\item{coxph_vars}{A string(vector). The variables used in the Cox-PH model.}
}
\value{
An integer indicating the height of the plot
}
\description{
Get the height of the plot depending on number of predictor variables
}
\author{
Kira E. Detrois
}
