% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpt_ids.R
\name{get_endpt_ids}
\alias{get_endpt_ids}
\title{Gets the individuals that are part of each study}
\usage{
get_endpt_ids(endpt_indvs_mat, endpt, all_ids)
}
\arguments{
\item{endpt_indvs_mat}{A data.frame. Contains at least column \code{ID},
and the current endpt. The endpoint column
has either 1 or 0, depending on whether
the individual is part of the study for
this endpoint.}

\item{all_ids}{A character (vector). The ids of all individuals.}
}
\description{
This is mostly relevant for FinnGen where we have some overalp for
the PRS studies and training and testing sets. However, it could
also be relevant for the training and testing of the PheRS.
}
