% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_obj.R
\name{get_endpt_sd_hr_ggplot}
\alias{get_endpt_sd_hr_ggplot}
\title{Creates a ggplot object for endpoint specific HR plots}
\usage{
get_endpt_sd_hr_ggplot(coxph_hrs, study_setup, preds)
}
\arguments{
\item{coxph_hrs}{A tibble. The Cox-PH HR results.
Needs to at least contain the columns \code{ENDPOINT}, \code{HR},
\code{CI_NEG}, \code{CI_POS}, \code{VAR}}

\item{ana_details}{A list. A list of parameters for the analysis.
It should contain the following fields: \code{study_type},
\code{preds}, \code{exp_len}, \code{wash_len}, \code{obs_len}, \code{obs_end_date}.}

\item{sort_hrs}{A logical.  If \code{True}, the endpoints are sorted by HR.}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that shows the HRs of the endpoints.
The x-axis shows the HRs, the y-axis shows the endpoints.
}
\author{
Kira E. Detrois
}
