% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_file_name}
\alias{get_file_name}
\title{Creats the file name for the HR plots with continuous scores}
\usage{
get_file_name(ana_details, res_type)
}
\value{
A character. The file name.
}
\description{
Creats the file name for the HR plots with continuous scores
}
\author{
Kira E. Detrois
}
