% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_full_res_path}
\alias{get_full_res_path}
\title{Adds downsampling and filter information to results directory path}
\usage{
get_full_res_path(write_res, res_dir, down_fctr, study_type, score_type)
}
\description{
Adds downsampling and filter information to results directory path
}
