% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{get_group_labs}
\alias{get_group_labs}
\title{Creates the labels for each risk score group}
\usage{
get_group_labs(score_group_tbl)
}
\arguments{
\item{score_group_tbl}{A named numeric. The risk score cuts for the
different quantiles.}
}
\value{
A character. The group names.
}
\description{
Creates the labels for each risk score group
}
\author{
Kira E. Detrois
}
