% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{get_groups_numeric}
\alias{get_groups_numeric}
\title{Gets the percentage of each group character as a numeric value}
\usage{
get_groups_numeric(indv_score_groups)
}
\arguments{
\item{indv_score_groups}{A character. The risk groups of each
individual.}
}
\value{
A numeric. The unique risk groups as a numeric vector.
}
\description{
Gets the percentage of each group character as a numeric value
}
\author{
Kira E. Detrois
}
