% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idnv_preds.R
\name{get_indv_model_probs}
\alias{get_indv_model_probs}
\title{Get's the predicted probability of the endpoint for the current model}
\usage{
get_indv_model_probs(coxph_mdl, surv_ana, study_data, preds)
}
\arguments{
\item{coxph_mdl}{An S4 \code{coxph} object. The current coxph model.}

\item{surv_ana}{An S4 \code{surv_ana} object. The current survival analysis setup.}

\item{study_data}{A data.frame. The current study data.}
}
\value{
A tibble with the predicted probabilities of the endpoint for
each individual for the current model.
}
\description{
Get's the predicted probability of the endpoint for the current model
}
