% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indvs_score_groups.R
\name{get_indvs_score_groups}
\alias{get_indvs_score_groups}
\title{Gets the risk score groups for each individual based on a score cutoff
table}
\usage{
get_indvs_score_groups(score_data, score_type, score_group_tbl)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}

\item{score_type}{A character. The score type. At the moment the two
options are \code{CCI}, and \code{PRS}.}

\item{score_group_tbl}{A named numeric. The risk score cuts for the
different quantiles.}
}
\value{
A factor. The risk score group for each individual.
}
\description{
Gets the risk score groups for each individual based on a score cutoff
table
}
\author{
Kira E. Detrois
}
