% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{get_min_indvs_data}
\alias{get_min_indvs_data}
\title{Get's the Cox-PH model results for runs where each group has at
least 5 controls and cases in each group}
\usage{
get_min_indvs_data(coxph_mdl, study)
}
\arguments{
\item{coxph_mdl}{A \code{coxph} object. See function \link[survival:coxph]{survival::coxph}.}

\item{study}{An S4 \code{study} object. The current study.
See class definition \link[Istudy:study]{Istudy::study}.}
}
\value{
A list. The filtered results.
}
\description{
Get's the Cox-PH model results for runs where each group has at
least 5 controls and cases in each group
}
\author{
Kira E. Detrois
}
