% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_n_group.R
\name{get_n_group_cases}
\alias{get_n_group_cases}
\title{Gets the number of cases of a score group for a given endpoint}
\usage{
get_n_group_cases(pheno_score_data, group_col_name, groups, endpt)
}
\arguments{
\item{pheno_score_data}{A data.frame with at least columns
SCORE_GROUP and the columns defined
by the variabel \code{endpt}.}

\item{groups}{A character. The groups in the Cox-PH model.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}
}
\description{
Gets the number of cases of a score group for a given endpoint
}
\author{
Kira E. Detrois
}
