% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_group.R
\name{get_n_group_cntrls}
\alias{get_n_group_cntrls}
\title{Gets the number of controls of a score group for a given
endpoint}
\usage{
get_n_group_cntrls(pheno_score_data, groups, endpt)
}
\arguments{
\item{groups}{A character. The groups in the Cox-PH model.}
}
\description{
Gets the number of controls of a score group for a given
endpoint
}
\author{
Kira E. Detrois
}
