% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_group.R
\name{get_n_group_indvs}
\alias{get_n_group_indvs}
\title{Gets the number of cases or controls of a score group for a given
endpoint}
\usage{
get_n_group_indvs(pheno_score_data, groups, endpt, indv_type)
}
\arguments{
\item{groups}{A character. The groups in the Cox-PH model.}

\item{indv_type}{A character. Can either be \code{cases} or \code{cntrls}.}
}
\description{
Gets the number of cases or controls of a score group for a given
endpoint
}
\author{
Kira E. Detrois
}
