% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_study_cidx.R
\name{get_non_na_pred_rows}
\alias{get_non_na_pred_rows}
\title{Exclude rows with missing values from a data frame}
\usage{
get_non_na_pred_rows(surv_ana)
}
\value{
A data frame with rows containing missing values in the specified
columns excluded.
}
\description{
This function filters a data frame to exclude rows with missing values
in the specified columns.
}
\author{
Kira E. Detrois
}
