% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_data.R
\name{get_pheno_score_data}
\alias{get_pheno_score_data}
\title{Extract Eligibility Score Data}
\usage{
get_pheno_score_data(
  score_type,
  pheno_data,
  study_setup,
  endpts_indvs_mat,
  icd_data = NULL,
  atc_data = NULL,
  prs_data = NULL,
  phers_data = NULL,
  endpt = NULL,
  min_indvs = 5
)
}
\value{
The data.frame with the score data for all eligible individuals
under the study setup of the current survival analysis setup.
Otherwise returns NULL. Also returns NULL if there is something
wrong with the \code{score_data}.
}
\description{
Extract the score and phenotype data of eligible cases and controls for
the Cox-PH analysis of a given endpoint.
}
\details{
If score type includes \code{CCI}, \code{PRS}, \code{EI}, \code{PheRS}, or \code{MED}, it also
calls \link{preprocess_score_data} to generate a data frame with the required
score data, and then joins the resulting score data with the phenotypic data
on the eligible individuals, using the function \link{join_dfs}.

It filters the study data to include only endpoints with both
cases and controls that have at least a minimum number of individuals.
}
\examples{
get_pheno_score_data("CCI", pheno_data, icd_data, endpt="J10_ASTHMA")

}
\author{
Kira E. Detrois
}
