% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HRs.R
\name{get_plot_preds}
\alias{get_plot_preds}
\title{Get Plot Preds}
\usage{
get_plot_preds(plot_preds, score_type)
}
\arguments{
\item{plot_preds}{A characte. The predictors to use when plotting HR.}

\item{score_type}{A character. The score types used in the Cox-PH model.}
}
\value{
A character vector of predictors to use when plotting hazard ratios.
}
\description{
This function returns the predictors to use when plotting hazard ratios.
If \code{plot_preds} is not NULL, it is returned. Otherwise, \code{score_type} is returned.
}
\author{
Kira E. Detrois
}
