% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_preds_file_name}
\alias{get_preds_file_name}
\title{Renames predictors for nicer files names and creates file name}
\usage{
get_preds_file_name(preds)
}
\arguments{
\item{preds}{A string (vector). The predictors to rename.}
}
\value{
A string. The file name.
}
\description{
Renames predictors for nicer files names and creates file name
}
\author{
Kira E. Detrois
}
