% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{get_pretty_covs_string}
\alias{get_pretty_covs_string}
\title{Turns vector of covariates into a string for plots}
\usage{
get_pretty_covs_string(covs)
}
\arguments{
\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\description{
Turns vector of covariates into a string for plots
}
