% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_data.R
\name{get_prob_data}
\alias{get_prob_data}
\title{Gets the current endpoint ZIP probability columns}
\usage{
get_prob_data(score_data, endpt)
}
\arguments{
\item{score_data}{A data.frame. The ZIP code probability data.
Contains a column of individual IDs and
a column named according to the selected
endpoint.}

\item{endpt}{A string. The current endpoint.}
}
\value{
A tibble. The renamed and filtered score data.
}
\description{
For the ZIP data the score columns have names in the
form of \code{J10_ASTHMA_ZIPprobs}. The function renames the
current ZIP column of interest to \code{ZIP}.
}
\author{
Kira E. Detrois
}
