% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_data.R
\name{get_prs_endpt_data}
\alias{get_prs_endpt_data}
\title{Gets the current endpoint PRS score column}
\usage{
get_prs_endpt_data(score_data, endpt)
}
\arguments{
\item{score_data}{A data.frame. The PRS data.
Contains a column of individual IDs and
a column named according to the selected
endpoint.}

\item{endpt}{A string. The current endpoint.}
}
\value{
A tibble. The renamed and filtered score data.
}
\description{
For the PRS data the score columns have names in the
form of \code{J10_ASTHMA_PRS}. The function renames the
current PRS column of interest to \code{PRS}.
}
\author{
Kira E. Detrois
}
