% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_data.R
\name{get_relevant_pheno_data_cols}
\alias{get_relevant_pheno_data_cols}
\title{Selects relevant columns from the phenotype data}
\usage{
get_relevant_pheno_data_cols(pheno_data, endpt)
}
\arguments{
\item{pheno_data}{A data.frame. The phenotype data.}

\item{endpt}{A string. The current endpoint.}
}
\value{
A tibble the relevant columns from the phenotype data.
\itemize{
\item \code{ID}
\item \code{SEX}
\item \code{DATE_OF_BIRTH}
\item \code{ANCESTRY}
\item The current endpoint i.e. \code{J10_ASTHMA}
\item The date for the current endpoint i.e. \code{J10_ASTHMA_DATE}
\item \code{END_OF_FOLLOWUP}
\item All principal components
\item \code{BATCH}
\item \code{ISCED_2011}
}
}
\description{
To decrease the size of downstream joins we do not
need all the original columns.
}
\author{
Kira E. Detrois
}
