% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_score_cut_file_name}
\alias{get_score_cut_file_name}
\title{Creats the file name for the score cut log file}
\usage{
get_score_cut_file_name(study, score_type, obs_end = NULL)
}
\arguments{
\item{study}{An S4 class representing the study setup.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}
}
\value{
A character. The file name.
}
\description{
Creats the file name for the score cut log file
}
\author{
Kira E. Detrois
}
