% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_group_tbl.R
\name{get_score_group_tbl}
\alias{get_score_group_tbl}
\title{Get's the score cutoff values}
\usage{
get_score_group_tbl(score_data, quantiles)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}

\item{quantiles}{A numeric. The probabilities for the quantiles.}
}
\value{
A named numeric vector with the quantiles as names
and score-cutoffs as values.
}
\description{
Get's the score cutoff values
}
\author{
Kira E. Detrois
}
