% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{get_score_groups}
\alias{get_score_groups}
\title{Get's the score cutoff values for the given probabilites.}
\usage{
get_score_groups(pheno_score_data)
}
\value{
A
}
\description{
Get's the score cutoff values for the given probabilites.
}
