% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_score_type_dir_name}
\alias{get_score_type_dir_name}
\title{Turns the score types and results description into a directory name}
\usage{
get_score_type_dir_name(score_type, res_descr = NULL)
}
\arguments{
\item{score_type}{A string (vector). The score types used in the analysis.}

\item{res_descr}{A string. An addition to add to the results directory.}
}
\value{
A string. The directory name.
}
\description{
Turns the score types and results description into a directory name
}
\author{
Kira E. Detrois
}
