% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_sd_title}
\alias{get_sd_title}
\title{Gets the title for a plot of HRs for each 1-SD increment}
\usage{
get_sd_title(coxph_vars)
}
\arguments{
\item{coxph_vars}{A character(vector). The variables used in the Cox-PH model.}
}
\value{
The title of the plot.
}
\description{
Takes the list of variables used in the Cox-PH model, and
depending on the number of unique variables returns a string
"1-SD Increment" or "\code{Variable_name} 1-SD Increment".
}
\author{
Kira E. Detrois
}
