% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_data.R
\name{get_study_cci_data}
\alias{get_study_cci_data}
\title{Gets the CCI score on the exposure window from ICD data}
\usage{
get_study_cci_data(pheno_data, icd_data, score_type, study_setup)
}
\arguments{
\item{pheno_data}{A data.frame. The phenotype data. Needs at least columns \code{ID}
and \code{DATE_OF_BIRTH}.}

\item{icd_data}{A data.frame. The ICD diagnoses codes. Needs at least columns
\code{ID}, \code{Event_age}, and \code{PRIMARY_ICD}.}

\item{score_type}{A string. Can be either \code{charlson} or \code{elixhauser}.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}
}
\value{
A tibble with columns \code{ID} and \code{CCI}, or \code{EI} depending on the
\code{score_type} used. Contains the charlson weighted comorbidity
scores for each individual.
}
\description{
Gets the CCI score on the exposure window from ICD data
}
\author{
Kira E. Detrois
}
