% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_cci_scores.R
\name{get_study_cci_scores}
\alias{get_study_cci_scores}
\title{Calcualtes the CCI score on the exposure window}
\usage{
get_study_cci_scores(elig_indv, icd_data)
}
\arguments{
\item{elig_indv}{A data.frame. The individuals which were eligble under
the current study setup. Needs to at least
contain the column defined in \code{ID}, and
\code{EXP_LEN} or \code{EXP_END}.}

\item{icd_data}{A data.frame with at least columns \code{ID}, \code{Event_age},
and \code{PRIMARY_ICD}.}
}
\value{
A data.frame with columns \code{ID}, and \code{SCORE}.
}
\description{
Calcualtes the CCI score on the exposure window
}
\author{
Kira E. Detrois
}
