% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_edu_data.R
\name{get_study_edu_data}
\alias{get_study_edu_data}
\title{Maps ISCED 2011 codes to age modes}
\usage{
get_study_edu_data(study_data)
}
\arguments{
\item{study_data}{A data frame containing columns \code{ID} and \code{ISCED_2011}}
}
\value{
A tibble containing columns \code{ID} and \code{EDU}
}
\description{
Maps the ISCED 2011 standard codes to the age modes better suited for the
downstream analysis by joining with a provided mapping data and renaming
the variable. Returns a single variable data frame with ID and the mapped
education level average ages.
}
\details{
The average ages are based on FinnGen R10. For more details see:
\verb{/finngen/red/detrois/scoio_processed/} in the SES Sandbox.
}
\examples{

study_data <- tibble::tibble(ID = 1:7, 
                             ISCED_2011 = c("1","2","3","4","5","6","7"))
get_study_edu_data(study_data)

}
\references{
The International Standard Classification of Education (ISCED) 2011
is a framework developed by UNESCO for classifying and comparing education
levels and programmes. For more information see:
\href{https://uis.unesco.org/en/topic/international-standard-classification-education-isced}{uis.onesco.org}
}
\author{
Kira E. Detrois
}
