% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_data.R
\name{get_study_med_data}
\alias{get_study_med_data}
\title{Gets the CCI score on the exposure window from ICD data}
\usage{
get_study_med_data(pheno_data, study_setup, atc_data)
}
\arguments{
\item{pheno_data}{A data.frame. The phenotype data. Needs at least columns \code{ID}
and \code{DATE_OF_BIRTH}.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}

\item{atc_data}{A data.frame. The ATC codes. Needs at least columns
\code{ID}, \code{ATC}, and \code{WEIGHT}.}
}
\value{
A tibble with columns \code{ID} and \code{MED}. Contains the weighted medication
scores for each individual.
}
\description{
Gets the CCI score on the exposure window from ICD data
}
\author{
Kira E. Detrois
}
