% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_study_subtitle}
\alias{get_study_subtitle}
\title{Creates a subtitle for a plot based on the current study setup}
\usage{
get_study_subtitle(study_setup)
}
\arguments{
\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}
}
\value{
A string. The subtitle describing the study setup.
}
\description{
Creates a subtitle for a plot based on the current study setup
}
\author{
Kira E. Detrois
}
