% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_surv_descr}
\alias{get_surv_descr}
\title{Returns the description of the survival analysis for use in plots}
\usage{
get_surv_descr(preds)
}
\arguments{
\item{preds}{A string (vector). The predictor variables
for the survival analysis}
}
\value{
A string with the description of the survival analysis
in a format suitable for use in plots.
}
\description{
Returns the description of the survival analysis for use in plots
}
\author{
Kira E. Detrois
}
