% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_ana.R
\name{get_surv_model_formula}
\alias{get_surv_model_formula}
\title{Creates the forumla string for the Cox-PH model.}
\usage{
get_surv_model_formula(endpt, covs)
}
\arguments{
\item{endpt}{A character. The column name of the current endpoint of
interest.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\value{
A character. The formula string for the Cox-PH model.
}
\description{
Creates the forumla string for the Cox-PH model.
}
