% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{get_surv_obj}
\alias{get_surv_obj}
\title{Creates a survival object for the selected endpoint}
\usage{
get_surv_obj(study_data, endpt)
}
\arguments{
\item{study_data}{A data.frame with all eligible individuals under
the chosen study setup and their scores. Needs at
least the columns defines in \code{covs}.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}
}
\description{
Creates a survival object for the selected endpoint. With time
being the age in years until onset or the age at the end of the
observation period.
}
\author{
Kira E. Detrois
}
