% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_two_level_groups.R
\name{get_two_level_groups}
\alias{get_two_level_groups}
\title{Gets the risk score groups for each individual based on a single
cutoff}
\usage{
get_two_level_groups(score_data, cutoff)
}
\arguments{
\item{score_data}{A data.frame. If \code{score_type = CCI} needs at least
column \code{SCORE}. If \code{score_type = PRS} columns of
form\code{J10_ASTHMA_PRS} depending on the chosen \code{endpts}.}

\item{cutoff}{A numeric. The score cutoff value.}
}
\value{
A character. The group names.
}
\description{
Gets the risk score groups for each individual based on a single
cutoff
}
\author{
Kira E. Detrois
}
