% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{get_unique_score_groups}
\alias{get_unique_score_groups}
\title{Get's the score cutoff values with empty quantiles removed}
\usage{
get_unique_score_groups(score_data, quantiles)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}

\item{quantiles}{A numeric. The probabilities for the quantiles.}
}
\value{
A named numeric vector with the quantiles as names
and score-cutoffs as values.
}
\description{
Get's the score cutoff values with empty quantiles removed
}
\author{
Kira E. Detrois
}
