% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_adds.R
\name{ggplot_add_points_and_errors}
\alias{ggplot_add_points_and_errors}
\title{Add points and error bars to a ggplot}
\usage{
ggplot_add_points_and_errors(plt, coxph_vars, study_type)
}
\arguments{
\item{plt}{The ggplot object to add points and error bars to.}

\item{coxph_vars}{A character(vector). The variables used in the Cox-PH model.}

\item{study_type}{A character. Can be either \code{forward} or \code{backward}.
\code{forward} considers individuals of a certain age
and a set exposure, washout, and observation periods
calcualted onwards from this age. It can simply be created
by creating a S4 study object.
\code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point. The exposure period will
be different for each individual depending on their birth
date.}
}
\value{
The modified ggplot object with points and error bars added.
}
\description{
Add points and error bars to a ggplot
}
