% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ana.R
\name{initialize,surv_ana-method}
\alias{initialize,surv_ana-method}
\title{Constructor for the \code{surv_ana} class}
\usage{
\S4method{initialize}{surv_ana}(.Object, ...)
}
\description{
Constructor for the \code{surv_ana} class
}
