% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dfs.R
\name{join_dfs}
\alias{join_dfs}
\title{Joins the phenotypic and score data}
\usage{
join_dfs(pheno_data, score_data, score_type = "CCI")
}
\arguments{
\item{pheno_data}{A data.frame with the phenotypic data on the individuals.
Should at least contain the columns:
\code{ID}, the columns specified in \code{covs} and
i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_AGE_FROM_BASE} where
the columns are the study endpoint and the age
from baseline, which will differ depending on the chosen
\code{endpts}.}

\item{score_type}{A character (vector).
The score types used in the Cox-PH model.}
}
\value{
A dataframe containing the merged study and score data.
}
\description{
Joins the dataframes \code{pheno_data}, and \code{score_data} using
the ID variable.
}
\details{
Sets missing values for the \code{CCI}, or \code{EI} to 0.
}
\author{
Kira E. Detrois
}
