% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{log_msg_table}
\alias{log_msg_table}
\title{Creates a string of the current study setup}
\usage{
log_msg_table(envir)
}
\arguments{
\item{envir}{A list with at least entry \code{pheno_score_data}.}
}
\value{
A tibble with columns \code{GROUP}, \code{SCORE_CUT}
}
\description{
Only used inside \code{write_log} function.
}
\author{
Kira E. Detrois
}
