% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{make_covs_fctrs}
\alias{make_covs_fctrs}
\title{Makes all character covariates a factor}
\usage{
make_covs_fctrs(elig_score_data, covs)
}
\arguments{
\item{elig_score_data}{A data.frame with all eligible individuals under
the chosen study setup and their scores. Needs at
least the columns defines in \code{covs}.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\description{
Makes all character covariates a factor
}
\author{
Kira E. Detrois
}
