% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HRs.R
\name{plot_age_sd_hr}
\alias{plot_age_sd_hr}
\title{Plots the HR from the continuous Cox-PH model of a forward study

Creates separate plots for each endpoint for the different age studies.}
\usage{
plot_age_sd_hr(surv_ana, coxph_hr_res, endpt, min_y, max_y)
}
\arguments{
\item{surv_ana}{An S4 class representing the current survival analysis
setup.}

\item{coxph_hr_res}{A tibble. The Cox-PH HR results. Needs
to at least contain the columns \code{ENDPOINT},
\code{HR}, \code{CI_NEG}, \code{CI_POS}, and \code{GROUP}.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}

\item{endpt_coxph_hr_res}{A tibble. The Cox-PH HR results. Needs
to at least contain the columns \code{ENDPOINT},
\code{HR}, \code{CI_NEG}, \code{CI_POS}, and \code{GROUP}.}
}
\description{
Plots the HR from the continuous Cox-PH model of a forward study

Creates separate plots for each endpoint for the different age studies.
}
\author{
Kira E. Detrois
}
