% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HRs.R
\name{plot_hrs}
\alias{plot_hrs}
\title{Plot the HR from the Cox-PH model}
\usage{
plot_hrs(coxph_hr_res, surv_ana)
}
\arguments{
\item{coxph_hr_res}{A tibble. The Cox-PH HR results. Needs
to at least contain the columns \code{ENDPOINT},
\code{HR}, \code{CI_NEG}, \code{CI_POS}, and \code{GROUP}.}

\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\description{
For forward study creates separate plots for each
endpoint for the different age studies. For
the backward study creates a single plot with the
different endpoints.
}
\author{
Kira E. Detrois
}
