% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HRs.R
\name{plot_risk_group_hr}
\alias{plot_risk_group_hr}
\title{Plots the HR from the risk stratified Cox-PH model of a forward study

Creates separate plots for each endpoint for the different age studies.
Plots the top 1\% compared to the 40-60\% group.}
\usage{
plot_risk_group_hr(surv_ana, endpt_coxph_hr_res, endpt)
}
\arguments{
\item{surv_ana}{An S4 class representing the current survival analysis
setup.}

\item{endpt_coxph_hr_res}{A tibble. The Cox-PH HR results. Needs
to at least contain the columns \code{ENDPOINT},
\code{HR}, \code{CI_NEG}, \code{CI_POS}, and \code{GROUP}.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}
}
\description{
Plots the HR from the risk stratified Cox-PH model of a forward study

Creates separate plots for each endpoint for the different age studies.
Plots the top 1\% compared to the 40-60\% group.
}
\author{
Kira E. Detrois
}
