% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_distr.R
\name{plot_score_distr}
\alias{plot_score_distr}
\title{Creates a plot of the score distribution}
\usage{
plot_score_distr(score_data, surv_ana)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}

\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\value{
The ggplot object.
}
\description{
Creates a plot of the score distribution
}
\author{
Kira E. Detrois
}
