% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_studies_score_distr.R
\name{plot_study_score_distr}
\alias{plot_study_score_distr}
\title{Calcualtes HR from a Cox-PH model for each endpoint}
\usage{
plot_study_score_distr(surv_ana)
}
\arguments{
\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\value{
A tibble with columns \code{Endpoint}, \code{Score}, \code{Group},
\code{N_controls}, \code{N_cases}, \code{beta}, \code{std_errs}, \code{p_val}, \code{HR},
\code{CI_pos}, \code{CI_neg}. The Cox-PH analysis results for all
selected endpoints.
}
\description{
Fits a Cox-proportional hazards model for the different endpt_studies.
}
\details{
Selects individuals from the data that are eligible given each of the
endpt_studies. For more details see package \code{Istudy} and function
\link[Istudy]{get_study_elig_indv}. Runs the analyis based on risk
groups. For \code{PRS} these are based on the quantiles, where the 40-60\%
group is the reference group. For \code{CCI} compares two groups based on a
cutoff defined in \code{bin_cut}.

For i.e. endpoint \code{J10_ASTHMA} and \code{covs = c("SEX", "YEAR_OF_BIRTH")}
the model would be
\code{Surv(J10_ASTHMA_AGE_DAYS, J10_ASTHMA) ~ SCORE_GROUP + SEX + YEAR_OF_BIRTH}.
}
\author{
Kira E. Detrois
}
