% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_score_data.R
\name{preprocess_score_data}
\alias{preprocess_score_data}
\title{Preprocess score data}
\usage{
preprocess_score_data(
  score_type,
  pheno_data,
  icd_data = NULL,
  atc_data = NULL,
  prs_data = NULL,
  phers_data = NULL,
  endpt = NULL
)
}
\arguments{
\item{score_type}{a character (vector). The types of scores to be preprocessed.
Available options include:
- "CCI" for Charlson comorbidity index,
- "EI" for Elixhauser index,
- "PRS" for polygenic risk score,
- "PheRS" for PheWAS results score,
- "MED" for medication data, and
- "EDU" for educational level. The column for this needs to be
called ISCED_2011 in the \code{pheno_data}.}

\item{pheno_data}{A dataframe containing phenotypic data.}

\item{icd_data}{A dataframe containing ICD data.
Required if "CCI" or "EI" is present in \code{score_type}}

\item{atc_data}{A dataframe containing ATC data.
Required if "MED" is present in \code{score_type}}

\item{prs_data}{A dataframe containing PRS data.
Required if "PRS" is present in \code{score_type}}

\item{phers_data}{A dataframe containing PheRS data.
Required if "PheRS" is present in \code{score_type}}

\item{endpt}{A character. The current endpoint.
Required if "PRS" or "PheRS" is present in \code{score_type}}
}
\value{
A dataframe containing all the preprocessed scores specified in
\code{score_type}.
}
\description{
Preprocesses different score data, based
on the selected scores and the current endpoint.
}
\author{
Kira E. Detrois
The education data comes from the phenotypic file
}
