% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{reformat_preds_pretty}
\alias{reformat_preds_pretty}
\title{Reformats predictors for plots}
\usage{
reformat_preds_pretty(preds)
}
\arguments{
\item{preds}{A string (vector).
The predictors to be reformatted.}
}
\value{
A string vector with the reformatted predictors.
}
\description{
Replaces certain string patterns by shorter and more pretty versions,
useful for axis labels, legends and captions.
}
\details{
The function also groups all PCs together if there are more than one.
}
\author{
Kira E. Detrois
}
