% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevel_to_mid_group.R
\name{relevel_to_mid_group}
\alias{relevel_to_mid_group}
\title{Change score group levels reference to middle}
\usage{
relevel_to_mid_group(indv_score_groups)
}
\arguments{
\item{indv_score_groups}{A character. The risk groups of each
individual.}
}
\value{
The risk group character vector releveled.
}
\description{
Relevel score groups to the 40-60\% group if it exists. Otherwise
sets the reference group is the group closest to 50\%.
}
\author{
Kira E. Detrois
}
