% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{relevel_to_middle}
\alias{relevel_to_middle}
\title{Change score group levels reference to middle}
\usage{
relevel_to_middle(indv_score_groups)
}
\arguments{
\item{indv_score_groups}{A character. The risk groups of each
individual.}
}
\value{
The risk group character vector releveled.
}
\description{
Relevel score groups so that the reference group is the one closest
to 50\%
}
\author{
Kira E. Detrois
}
