% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_age_exp_studies.R
\name{run_age_exp_endpt_studies}
\alias{run_age_exp_endpt_studies}
\title{Calcualtes HR from a Cox-PH model for each exposure age and endpoint}
\usage{
run_age_exp_endpt_studies(
  pheno_data,
  score_ages_data,
  score_type,
  endpts = c("J10_ASTHMA"),
  exp_ages = c(20, 30, 40),
  exp_len = 10,
  wash_len = 2,
  obs_len = 8,
  downsample_fctr = NULL,
  ancs = "EUR",
  covs = c("SEX", "YEAR_OF_BIRTH"),
  bin_cut = 1,
  min_indvs = 5,
  write_res = FALSE,
  res_dir = NULL
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, the columns specified in \code{covs} and
i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_AGE_DAYS} where
the columns are the study endpoint and date, which
will differ depending on the chosen \code{endpts}.}

\item{score_ages_data}{A named vector of tibbles.
The score data for each indiviuals for different
exposure periods.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{endpts}{A string. The column names of the endpoints of
interest.}

\item{exp_ages}{An integer (vector). Ages at which exposure period
starts (in years).}

\item{exp_len}{An integer. Length of the exposure period ^
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type == CCI}.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\description{
For more details see \link{calc_endpt_studies_hrs} function documentation.
}
\author{
Kira E. Detrois
}
