% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_ana.R
\name{run_and_add_coxph_ana}
\alias{run_and_add_coxph_ana}
\title{Fits the Cox model and adds the results to the data.frame \code{coxph_res}}
\usage{
run_and_add_coxph_ana(
  coxph_res,
  pheno_score_data,
  score_type,
  endpt,
  predictor
)
}
\arguments{
\item{predictor}{A character. The predictor for the survival.}
}
\value{
A tibble the results data.fram \code{coxph_res} with added
columns from the analysis run.
}
\description{
Fits the Cox model and adds the results to the data.frame \code{coxph_res}
}
