% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_ana.R
\name{run_coxph_ana}
\alias{run_coxph_ana}
\title{Fits the Cox model for a given endpoint and predictor}
\usage{
run_coxph_ana(pheno_score_data, endpt, covs)
}
\arguments{
\item{pheno_score_data}{A data.frame with at least the columns:
\code{ID}, \code{SCORE_GROUP}, the columns specified in \code{covs}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE} where
the columns are the study endpoint and date, which
will differ depending on the input variable \code{endpts}.}

\item{endpt}{A string. The column name of the current endpoint of
interest.}

\item{covs}{A character. The column names of the covariates to add to
the predictor of the Cox-PH model.}
}
\description{
Fits the Cox model for a given endpoint and predictor
}
\author{
Kira E. Detrois
}
