% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv_studies.R
\name{run_surv_studies}
\alias{run_surv_studies}
\title{Calcualtes HR from a Cox-PH model for each exposure age and endpoint}
\usage{
run_surv_studies(
  pheno_data,
  icd_data = NULL,
  prs_data = NULL,
  score_type,
  study_type = "forward",
  endpts = c("J10_ASTHMA"),
  exp_ages = 0,
  exp_len = NULL,
  wash_len = 2,
  obs_len = 8,
  obs_end_date = as.Date("2021/01/01"),
  downsample_fctr = NA_integer_,
  ancs = NA_character_,
  max_age = 90,
  filter_1998 = FALSE,
  covs = c("SEX", "YEAR_OF_BIRTH"),
  bin_cut = 1,
  min_indvs = 5,
  write_res = FALSE,
  res_dir = NULL
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, the columns specified in \code{covs} and
i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_AGE_DAYS} where
the columns are the study endpoint and date, which
will differ depending on the chosen \code{endpts}.
The phenotypic data on the individuals.}

\item{score_type}{A character. The score type. At the moment the two
options are \code{CCI}, and \code{PRS}.}

\item{study_type}{A character. Can be either \code{forward} or \code{backward}.
\code{forward} considers individuals of a certain age
and a set exposure, washout, and observation periods
calcualted onwards from this age. It can simply be created
by creating a S4 study object.
\code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point. The exposure period will
be different for each individual depending on their birth
date. This setup can be created, using the function
\code{\link{get_backward_study}}.}

\item{endpts}{A character (vector). The column names of the endpoints of
interest.}

\item{exp_ages}{An integer (vector). Ages at which exposure period
starts (in years).}

\item{exp_len}{An integer. Length of the exposure period ^
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}

\item{obs_end_date}{A Date. The end of the observation period. Needed
for \code{backward} studies. If not set the end will
be the most recent date across the diagnosis columns.}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{ancs}{A character (vector). The ancestries to consider.}

\item{max_age}{A numeric. The maximum age at the end of the exposure
window of individuals.}

\item{filter_1998}{A boolean. Whether to filter out individuals
where the exposure period is before 1998.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type == CCI}.}

\item{min_indvs}{An integer. The minimum number of individuals each
group in the analyses needs to have. This is important
for being able to export data from the FinnGen Sanbdox.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}

\item{score_data}{A data.frame with at least columns \code{SCORE}, or
i.e. \code{J10_ASTHMA_PRS} for endpoint \code{J10_ASTHMA}
and score type \code{PRS}.}
}
\description{
Calcualtes HR from a Cox-PH model for each exposure age and endpoint
}
\author{
Kira E. Detrois
}
