% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_save.R
\name{save_plt}
\alias{save_plt}
\title{Saves a ggplot object to a file}
\usage{
save_plt(file_path, plt, width, height)
}
\arguments{
\item{file_path}{A character. The file path to save the plot to.}

\item{plt}{ggplot object. The plot to be saved.}

\item{width}{numeric. The width of the image file.}

\item{height}{numeric. The height of the image file.}
}
\description{
Save a ggplot object as an png image file
}
\author{
Kira E. Detrois
}
