% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv_studies.R
\name{save_results}
\alias{save_results}
\title{Saves the results}
\usage{
save_results(hr_res, cidx_res, study, surv_ana)
}
\arguments{
\item{hr_res}{A tibble containing the hazard ratios for each exposure age.}

\item{cidx_res}{A tibble containing the c-index for each exposure age.}

\item{surv_ana}{A survival analysis object}
}
\description{
This function saves the results of a survival analysis and plots them
Saves everything to files if this is wanted.
}
