% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ana.R
\name{setPreds,surv_ana-method}
\alias{setPreds,surv_ana-method}
\title{Sets \code{endpt} of the S4 study object to a new value}
\usage{
\S4method{setPreds}{surv_ana}(.Object, crnt_score_types)
}
\arguments{
\item{.Object}{The S4 study object.}

\item{endpt}{A character. The new endpoint.}
}
\description{
Sets \code{endpt} of the S4 study object to a new value
}
