% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ana.R
\name{setPreds}
\alias{setPreds}
\title{Sets \code{endpt} of the S4 study object to a new value}
\usage{
setPreds(.Object, crnt_score_types)
}
\arguments{
\item{.Object}{The S4 surv_ana object.}

\item{preds}{A character. The new predictors.}

\item{plot_preds}{A character. The new predictors for potting}
}
\description{
Sets \code{endpt} of the S4 study object to a new value
}
