% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coxph_mdl.R
\name{update_coxph_log}
\alias{update_coxph_log}
\title{Adding new zph analysis of the updated coxph model with interaction terms to the log}
\usage{
update_coxph_log(coxph_mdl, study, surv_ana)
}
\arguments{
\item{coxph_mdl}{The updated coxph model}

\item{study}{The study object}

\item{surv_ana}{The survival analysis object}
}
\description{
Adding new zph analysis of the updated coxph model with interaction terms to the log
}
