% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_log.R
\name{write_coxph_to_log}
\alias{write_coxph_to_log}
\title{writing zph results of coxph model to log}
\usage{
write_coxph_to_log(coxph_mdl, study, surv_ana)
}
\arguments{
\item{coxph_mdl}{The coxph model}

\item{study}{The study object}

\item{surv_ana}{The survival analysis object}
}
\description{
writing zph results of coxph model to log
}
