% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_hr_res}
\alias{write_hr_res}
\title{Writes model results to a tab-delim file}
\usage{
write_hr_res(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{write_hr_res}, \code{res_dir} ,
\code{all_coxph_res}, \code{endpt}, \code{exp_age}, \code{exp_len}, \code{wash_len},
and \code{obs_len}.}
}
\description{
Writes model results to a tab-delim file
}
\author{
Kira E. Detrois
}
