% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_hr_res_file}
\alias{write_hr_res_file}
\title{Writes model results to a tab-delim file}
\usage{
write_hr_res_file(
  endpt_hrs_tib,
  score_type,
  bin_cut,
  study,
  write_res,
  res_dir
)
}
\arguments{
\item{endpt_hrs_tib}{A tibble with the results for previous endpts.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type == CCI}.}

\item{study}{An S4 class representing the study setup.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\description{
Writes model results to a tab-delim file
}
\author{
Kira E. Detrois
}
