% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_res_file}
\alias{write_res_file}
\title{Writes model results to a tab-delim file}
\usage{
write_res_file(coxph_res, study, write_res, res_dir)
}
\description{
Writes model results to a tab-delim file
}
\author{
Kira E. Detrois
}
