% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_score_groups_to_log}
\alias{write_score_groups_to_log}
\title{Write score cut-off values and groups to log}
\usage{
write_score_groups_to_log(score_group_tbl, study, write_res, res_dir)
}
\description{
Write score cut-off values and groups to log
}
\author{
Kira E. Detrois
}
