% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ana.R
\name{write_to_error_file}
\alias{write_to_error_file}
\title{Writes error message to file if path is not Null or NA}
\usage{
write_to_error_file(error_file, msg)
}
\arguments{
\item{error_file}{A character. The path to the error file.}

\item{msg}{A character. The message to write to the error file.}
}
\description{
Writes error message to file if path is not Null or NA
}
\author{
Kira E. Detrois
}
