% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{add_cidx_res_row}
\alias{add_cidx_res_row}
\title{Adds a row to the C-index results tibble}
\usage{
add_cidx_res_row(c_idxs_res, c_idx_res, surv_ana, study)
}
\arguments{
\item{c_idxs_res}{A tibble with the results for previous endpts.}

\item{c_idx_res}{A tibble with the new results to add.}

\item{surv_ana}{An S4 \code{surv_ana} object. The current survival analysis setup.
See class definition \link{surv_ana}.}

\item{study}{An S4 \code{study} object. The current study.
See class definition \link[Istudy:study]{Istudy::study}.}
}
\value{
A tibble. The updated \code{c_idxs_res} with the added results for
the current endpoint.
}
\description{
Adds a row to the C-index results tibble
}
\author{
Kira E. Detrois
}
