% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{add_cidx_res_row}
\alias{add_cidx_res_row}
\title{Adds a row to the C-index results tibble}
\usage{
add_cidx_res_row(endpt_c_idxs_tib, c_idx_res, surv_ana)
}
\arguments{
\item{endpt_c_idxs_tib}{A tibble with the results for previous endpts.}

\item{c_idx_res}{A tibble with the new results to add.}
}
\value{
A tibble. The updated \code{endpt_c_idxs_tib} with the added results for
the current endpoint.
}
\description{
Adds a row to the C-index results tibble
}
\author{
Kira E. Detrois
}
