% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coxph_res_row.R
\name{add_coxph_res_row}
\alias{add_coxph_res_row}
\title{Adds a row to the results tibble}
\usage{
add_coxph_res_row(coxph_res_tib, coxph_mdl, score_type, endpt, elig_indv)
}
\arguments{
\item{coxph_res_tib}{A tibble with the results for previous endpoints.}

\item{coxph_mdl}{A Cox-PH model.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}

\item{elig_indv}{A tibble. The individuals which were eligble under
the current study setup. Needs to at least
contain the column defined in \code{endpt}.}
}
\value{
A tibble. The updated \code{coxph_res_tib} with the added results for
the current endpoint.
}
\description{
Adds a row to the results tibble
}
\author{
Kira E. Detrois
}
