% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{add_risk_group_col}
\alias{add_risk_group_col}
\title{Adds a risk group column}
\usage{
add_risk_group_col(
  score_data,
  score_type,
  study,
  bin_cut = NA_integer_,
  write_res = FALSE,
  res_dir = NA_character_
)
}
\arguments{
\item{score_data}{A data.frame. Needs at least column \code{SCORE} if
\code{score_type = CCI}, or i.e. \code{J10_ASTHMA_PRS} for
\code{score_type = PRS} depending on the chosen \code{endpts}.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{study}{An S4 class representing the study setup.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type} is \code{CCI}.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\value{
The data.frame with the added risk group column
}
\description{
The risk groups are quantiles based on the score distribution
of all individuals in the data. They are given as left-open
intervals i.e. first risk group could be an interval 0\% <= x < 60\%
and would be called \verb{Group 60\%}, the second risk group could then
go from i.e. 60\% <= x <= 80\%  and would be indicated as \verb{Group 80\%}
etc.
}
\author{
Kira E. Detrois
}
