% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{add_risk_group_col}
\alias{add_risk_group_col}
\title{Adds a risk group column}
\usage{
add_risk_group_col(
  score_data,
  score_type,
  study,
  bin_cut = 1,
  write_res = FALSE,
  res_dir = NULL
)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{study}{An S4 class representing the study setup.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type == CCI}.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\value{
The data.frame with the added risk group column
}
\description{
For CCI gets two risk groups based on the cutoff value \code{bin_cut}.
The groups are then \verb{<= bin_cut} and \verb{>bin_cut}.
}
\details{
For PRS bases the risk grouping on the quantiles
\code{c(0,0.01,0.05,0.1,0.2,0.4,0.6,0.8,0.9,0.95,0.99,1)} for the score
distribution of all individuals in the data.
The resulting intervals are left-open, except the first one which
is both right- and left-open. I.e. the first risk group is an
interval 0\% <= x <= 1\% and the second risk group is then the interval
1\% < x <= 5\% etc.
}
\author{
Kira E. Detrois
}
