% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{add_risk_group_col}
\alias{add_risk_group_col}
\title{Adds a risk group column}
\usage{
add_risk_group_col(pheno_score_data)
}
\value{
The data.frame with the added risk group column
}
\description{
The risk groups are quantiles based on the score distribution
of all individuals in the data. They are given as left-open
intervals I.e. first risk group could be an interval 0\% <= x < 60\%
and would be called \verb{Group 60\%}, the second risk group could then
go from i.e. 60\% <= x <= 80\%  and would be indicated as \verb{Group 80\%}
etc.
}
\author{
Kira E. Detrois
}
