% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_hrs.R
\name{calc_endpt_hrs}
\alias{calc_endpt_hrs}
\title{Calcualtes HR from a cox-ph model for each endpoint}
\usage{
calc_endpt_hrs(
  pheno_data,
  score_data,
  score_col_name,
  score_type,
  endpts,
  exp_age = 30,
  exp_len = 10,
  wash_len = 2,
  obs_len = 8,
  downsample_fctr = NA,
  write_res = FALSE,
  res_dir = NA
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpts}.}

\item{score_data}{A data.frame with the score results for each individuals.
Should have at least column defined in \code{score_col_name}.}

\item{score_col_name}{A character. The name of the column with the scores.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{endpts}{A string. The column names of the endpoints of
interest.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\description{
Calcualtes HR from a cox-ph model for each endpoint
}
\author{
Kira E. Detrois
}
