% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_study_cidx.R
\name{calc_endpt_study_cidx}
\alias{calc_endpt_study_cidx}
\title{Calcualtes HR from a Cox-PH model for the current survival analysis
setup}
\usage{
calc_endpt_study_cidx(surv_ana)
}
\arguments{
\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\value{
A tibble with columns \code{Endpoint}, \code{Score}, \code{Group},
\code{N_controls}, \code{N_cases}, \code{beta}, \code{std_errs}, \code{p_val}, \code{HR},
\code{CI_pos}, \code{CI_neg}. The Cox-PH analysis results for all
selected endpoints.
}
\description{
For i.e. a study with endpoint \code{J10_ASTHMA} and
\code{covs = c("SEX", "YEAR_OF_BIRTH")} the model would be
\code{Surv(J10_ASTHMA_AGE_DAYS, J10_ASTHMA) ~ SCORE + SEX + YEAR_OF_BIRTH}.
}
\author{
Kira E. Detrois
}
