% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_studies_hrs.R
\name{calc_studies_hrs}
\alias{calc_studies_hrs}
\title{Calcualtes HR from a Cox-PH model for each endpoint}
\usage{
calc_studies_hrs(
  pheno_data,
  score_data,
  score_col_name,
  score_type,
  studies,
  write_res = FALSE,
  res_dir = NA
)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpts}.}

\item{score_data}{A data.frame. The score results for each individual.
Should have at least column defined in
\code{score_col_name} and column \code{ID}.}

\item{score_col_name}{A character. The name of}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{studies}{A vector of S4 classes representing the study setups.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\description{
Calcualtes HR from a Cox-PH model for each endpoint
}
\author{
Kira E. Detrois
}
