% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{check_and_get_file_path}
\alias{check_and_get_file_path}
\title{Creats the file directory and name for the different results}
\usage{
check_and_get_file_path(
  score_type,
  study,
  write_res,
  res_dir,
  res_type,
  bin_cut = 1
)
}
\arguments{
\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{study}{An S4 class representing the study setup.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}

\item{res_type}{A character. The results type.}

\item{bin_cut}{A numeric. The binary cutoff value for classifying high
and low score individuals. Currently only in use if
the \code{score_type == CCI}.}
}
\value{
A character. The file name.
}
\description{
Creats the file directory and name for the different results
}
\author{
Kira E. Detrois
}
