% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_endpt_study_obj.R
\name{create_endpt_study_obj}
\alias{create_endpt_study_obj}
\title{Creates an S4 study object}
\usage{
create_endpt_study_obj(
  study_data,
  study_type = "forward",
  preds = "CCI",
  endpt = "J10_ASTHMA",
  exp_age = NULL,
  exp_len = NULL,
  wash_len = 2,
  obs_len = 8,
  obs_end_date = as.Date("2021/01/01"),
  down_fctr = NA_integer_,
  ancs = c("EAS"),
  max_age = 200,
  write_res = FALSE,
  res_dir = NA_character_
)
}
\arguments{
\item{study_type}{A character. Can be either \code{forward} or \code{backward}.
\code{forward} considers individuals of a certain age
and a set exposure, washout, and observation periods
calcualted onwards from this age. It can simply be created
by creating a S4 study object.
\code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point. The exposure period will
be different for each individual depending on their birth
date. This setup can be created, using the function
\code{\link{get_backward_study}}.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years).}

\item{exp_len}{An integer. Length of the exposure period ^
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}

\item{obs_end_date}{A Date. The end of the observation period. Needed
for \code{backward} studies. If not set the end will
be the most recent date across the diagnosis columns.}

\item{down_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{ancs}{A character (vector). The ancestries to consider.}

\item{max_age}{A numeric. The maximum age at the end of the exposure
window of individuals.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\value{
An S4 study object.
}
\description{
Creates an S4 study object
}
\author{
Kira E. Detrois
}
