% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coxph_res.R
\name{extract_coxph_res}
\alias{extract_coxph_res}
\title{Extract relevant results from the Cox-PH model}
\usage{
extract_coxph_res(coxph_mdl, score_type)
}
\arguments{
\item{coxph_mdl}{A \code{coxph} object. See function \link[survival:coxph]{survival::coxph}.}

\item{score_type}{A string (vector). The score types used in the analysis.}
}
\value{
A list with elements
\itemize{
\item beta: Estimated regression coefficients
\item std_err: Standard errors of the coefficients
\item p_val: P-values of the coefficients
\item HR: Hazard ratios
\item CI_neg: Negative confidence interval limits for the hazard ratios
\item CI_pos: Positive confidence interval limits for the hazard ratios
\item groups: Groupings for the predictors
\item preds: Predictor names
}
If the model is NULL, an empty list is returned
}
\description{
This function extracts the results of a Cox proportional hazards model
fit and returns them in a list.
}
\author{
Kira E. Detrois
}
