% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coxph_res.R
\name{extract_coxph_res}
\alias{extract_coxph_res}
\title{Extracts the relevant results from the Cox-PH model}
\usage{
extract_coxph_res(coxph_mdl)
}
\arguments{
\item{coxph_mdl}{A Cox-PH model.}
}
\value{
A list(\code{beta}, \code{std_err}, \code{p_val}, \code{HR}, \code{CI}, \code{groups}).
}
\description{
Extracts the relevant results from the Cox-PH model
}
\author{
Kira E. Detrois
}
