% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{filter_out_missing_hrs}
\alias{filter_out_missing_hrs}
\title{Filters out HRs that are NA, or where either bound of the CI is infinite}
\usage{
filter_out_missing_hrs(coxph_hrs)
}
\arguments{
\item{coxph_hrs}{A tibble. The Cox-PH HR results. Needs to at least contain
the columns \code{HR}, \code{CI_NEG}, and \code{CI_POS}.}
}
\value{
A tibble with HRs that are not missing and where neither the upper
or lower bound of the CI is infinite.
}
\description{
Filters out HRs that are NA, or where either bound of the CI is infinite
}
\author{
Kira E. Detrois.
}
