% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_obj.R
\name{get_age_sd_hr_ggplot}
\alias{get_age_sd_hr_ggplot}
\title{Creats ggplot for forwar study setup}
\usage{
get_age_sd_hr_ggplot(coxph_hrs, study_setup, preds, endpt, min_y, max_y)
}
\arguments{
\item{coxph_hrs}{A data.frame. The results of the Cox-PH model for
the current endpoint.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}

\item{preds}{A string (vector). The predictors of the analysis
for the caption.}

\item{endpt}{A string. The current endpoint.}

\item{min_y}{A numeric. Minimum value for the y-axis.}

\item{max_y}{A numeric. Maximum value for the y-axis.}
}
\value{
A ggplot object.
}
\description{
Creates the ggplot for the HRs vs. observation age for an endpoint
in a forward study.
}
\author{
Kira E. Detrois
}
