% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_all_preds_sorted}
\alias{get_all_preds_sorted}
\title{Get All Preds Sorted}
\usage{
get_all_preds_sorted(score_type, covs = NULL)
}
\arguments{
\item{score_type}{A string (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index,
\item \code{EI} for Elixhauser index,
\item \code{PRS} for polygenic risk score,
\item \code{PheRS} for PheRS results score,
\item \code{MED} for medication data, and
\item \code{EDU} for educational level. The column for this needs to
be called \code{ISCED_2011} in the \code{pheno_data}.
\item \code{ZIP} for zip scores,
}}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\value{
A string vector of all predictors being used in the survival analysis,
sorted by predictor name.
}
\description{
Returns a sorted character vector of all predictors being used in the survival
analysis. Interaction predictors are sorted separately from non-interaction
predictors.
}
\author{
Kira E. Detrois
}
