% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_caption}
\alias{get_caption}
\title{Returns the plot caption for a HRs plot}
\usage{
get_caption(study_setup, preds)
}
\arguments{
\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}

\item{preds}{A character (vector). The predictors of the analysis.}
}
\value{
A string. The plot caption.
}
\description{
The function generates a caption that describes the study setup
and the predictor variables used in the Cox-PH model.
}
\author{
Kira E. Detrois
}
