% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_CI}
\alias{get_CI}
\title{95\% confidence interval given the ML estimator and SE}
\usage{
get_CI(ML, SE)
}
\arguments{
\item{ML}{maximum likelihood estimator of the parameter}

\item{SE}{standard error of the ML estimator}
}
\description{
95\% confidence interval given the ML estimator and SE
}
