% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{get_coxph_formula}
\alias{get_coxph_formula}
\title{Creates the forumla string for the Cox-PH model}
\usage{
get_coxph_formula(surv_ana, pred_score = "SCORE_GROUP", pec = FALSE)
}
\arguments{
\item{surv_ana}{An S4 class representing the current survival analysis
setup.}

\item{pred_score}{A character. The score type to predict. For direct
regression \code{SCORE}, and for regression on the
risk group \code{SCORE_GROUP}.}
}
\value{
A character. The formula string for the Cox-PH model.
}
\description{
Creates the forumla string for the Cox-PH model
}
