% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{get_coxph_formula}
\alias{get_coxph_formula}
\title{Creates the forumla string for the Cox-PH model}
\usage{
get_coxph_formula(surv_ana)
}
\value{
A character. The formula string for the Cox-PH model.
}
\description{
Creates the forumla string for the Cox-PH model
}
