% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coxph_mdl.R
\name{get_coxph_formula}
\alias{get_coxph_formula}
\title{Creates the forumla string for the Cox-PH model}
\usage{
get_coxph_formula(preds, endpt)
}
\arguments{
\item{preds}{A string (vector). The predictors for the model.}

\item{endpt}{A string. The current endpoint.}
}
\value{
The formula for the Cox-PH model.

A string. The formula string for the Cox-PH model.
}
\description{
Will be i.e.
\code{Surv(J10_ASTHMA_AGE_FOR_BASE, J10_ASTHMA) ~ SEX + YEAR_OF_BIRTH}
for endpoint \code{J10_ASTHMA} and predictors \code{SEX} and \code{YEAR_OF_BIRTH}.
}
