% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crnt_endpt_data.R
\name{get_crnt_endpt_data}
\alias{get_crnt_endpt_data}
\title{Gets the data of individuals for a given endpoint study}
\usage{
get_crnt_endpt_data(endpts_indvs_mat = NULL, endpt, pheno_data)
}
\arguments{
\item{endpts_indvs_mat}{A dataframe. Contains a column of individual
IDs and a binary column for each endpoint.}

\item{endpt}{A character. The endpoint of interest,
must match a column name in \code{endpts_indvs_mat}}

\item{all_ids}{A character (vector). The IDs of all individuals.}
}
\value{
A data.frame that contains the data for individuals that can be used
for the current endpoint.
}
\description{
This is mostly relevant for FinnGen where we have some overalp for
the PRS studies and training and testing sets. However, it could
also be relevant for the training and testing of the PheRS.
}
\author{
Kira E. Detrois
}
