% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_down_dir}
\alias{get_down_dir}
\title{Creates directory name for the downsampling factor}
\usage{
get_down_dir(down_fctr)
}
\arguments{
\item{down_fctr}{A numeric. Defines how many controls there should be for every case.}
}
\value{
A character. The directory name describing the downsampling factor used.
}
\description{
If the downsampling factor is NA then the direcotry is
named \code{no_down}.
}
\author{
Kira E. Detrois
}
