% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elig_score_data.R
\name{get_elig_score_data}
\alias{get_elig_score_data}
\title{Gets the full score and phenotype data on the eligible individuals}
\usage{
get_elig_score_data(
  score_type,
  study_data,
  icd_data = NULL,
  atc_data = NULL,
  prs_data = NULL,
  endpt = NULL,
  min_indvs = 5
)
}
\value{
The data.frame with the score data for all eligible individuals
under the study setup of the current survival analysis setup.
}
\description{
Should be run before any analyis on the survival setup is run.
}
\details{
For \code{CCI} calculates the CCI score based on the ICD-data.
Also, renames the score column to \code{SCORE} and filters out NAs.
For the PRS data the score columsn have names in the
form of \code{J10_ASTHMA_PRS}. The function renames the
current PRS column of interest to \code{SCORE}. Then it
filters out all NAs in the column. See function
\code{\link{get_prs_endpt_scores}}.

Joins the resulting score data with the phenotypic data
on the eligible individuals.

Checks wether there are enough individuals in the case
and control groups. Otherwise returns an empty data.frame.
}
\author{
Kira E. Detrois
}
