% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_endpt_score_file_name}
\alias{get_endpt_score_file_name}
\title{Creats the file name for endpoint specific score distribution plot}
\usage{
get_endpt_score_file_name(surv_ana)
}
\arguments{
\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\value{
A character. The file name.
}
\description{
Creats the file name for endpoint specific score distribution plot
}
\author{
Kira E. Detrois
}
