% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_obj.R
\name{get_endpt_sd_hr_ggplot}
\alias{get_endpt_sd_hr_ggplot}
\title{Creates a ggplot object for endpoint specific HR plots}
\usage{
get_endpt_sd_hr_ggplot(coxph_hrs, study_setup, preds)
}
\arguments{
\item{coxph_hrs}{A tibble. The Cox-PH HR results.
Needs to at least contain the columns \code{ENDPOINT}, \code{HR},
\code{CI_NEG}, \code{CI_POS}, \code{VAR}}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}

\item{preds}{A string (vector). The predictors of the analysis.}
}
\value{
A ggplot object
}
\description{
Creates a ggplot object that shows the HRs of the endpoints.
The x-axis shows the HRs, the y-axis shows the endpoints.
}
\author{
Kira E. Detrois
}
