% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_group_col.R
\name{get_indvs_score_groups}
\alias{get_indvs_score_groups}
\title{Gets the risk score groups for each individual based on a score cutoff
table}
\usage{
get_indvs_score_groups(score_data)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}
}
\value{
A factor. The risk score group for each individual.
}
\description{
Gets the risk score groups for each individual based on a score cutoff
table
}
\author{
Kira E. Detrois
}
