% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{get_min_indvs_data}
\alias{get_min_indvs_data}
\title{Get's the Cox-PH model results for runs where each group has at
least 5 controls and cases in each group}
\usage{
get_min_indvs_data(coxph_mdl, surv_ana)
}
\arguments{
\item{coxph_mdl}{A Cox-PH model.}

\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\value{
A list. The filtered results.
}
\description{
Get's the Cox-PH model results for runs where each group has at
least 5 controls and cases in each group
}
\author{
Kira E. Detrois
}
