% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tibbles.R
\name{get_min_indvs_data}
\alias{get_min_indvs_data}
\title{Get's the Cox-PH model results for runs where each group has at
least 5 controls and cases in each group.}
\usage{
get_min_indvs_data(elig_indv, coxph_mdl, study, min_indvs = 5)
}
\arguments{
\item{elig_indv}{A tibble. The individuals which were eligble under
the current study setup. Needs to at least
contain the column defined in \code{endpt}.}

\item{coxph_mdl}{A Cox-PH model.}
}
\description{
Get's the Cox-PH model results for runs where each group has at
least 5 controls and cases in each group.
}
