% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_n_group.R
\name{get_n_group_indvs}
\alias{get_n_group_indvs}
\title{Gets the number of cases or controls of a score group for a given
endpoint}
\usage{
get_n_group_indvs(pheno_score_data, group_col_name, groups, endpt, indv_type)
}
\arguments{
\item{pheno_score_data}{A data.frame with at least columns
SCORE_GROUP and the columns defined
by the variabel \code{endpt}.}

\item{groups}{A character. The groups in the Cox-PH model.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}

\item{indv_type}{A character. Defines whether to get the cases
or controls for each group. Has to be
either \code{cntrls}, or \code{cases}.}
}
\description{
Gets the number of cases or controls of a score group for a given
endpoint
}
\author{
Kira E. Detrois
}
