% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_obs_age_period_str}
\alias{get_obs_age_period_str}
\title{Create strings for the age in the observation period for each group}
\usage{
get_obs_age_period_str(coxph_hrs, study_setup)
}
\arguments{
\item{coxph_hrs}{A tibble. The Cox-PH HR results.
Needs to at least contain the columns \code{EXP_AGE}-}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup-class]{Istudy::study_setup}.}
}
\value{
A string vector with the observation period strings
for each age group for the axis labels of the plot.
}
\description{
These are used as axis labels in the ggplot for the forward study setup
}
\author{
Kira E. Detrois
}
