% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_score_data.R
\name{get_phers_endpt_data}
\alias{get_phers_endpt_data}
\title{Renames the PRS score column to \code{PheRS}}
\usage{
get_phers_endpt_data(score_data, endpt)
}
\arguments{
\item{score_data}{A data.frame with at least columns \code{SCORE}, or
i.e. \code{J10_ASTHMA_PRS} for endpoint \code{J10_ASTHMA}
and score type \code{PRS}.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}
}
\value{
A tibble. The renamed and filtered score data.
}
\description{
For the PRS data the score columns have names in the
form of \code{J10_ASTHMA_PRS}. The function renames the
current PRS column of interest to \code{PheRS}.
}
