% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_plot_preds}
\alias{get_plot_preds}
\title{Get Plot Preds}
\usage{
get_plot_preds(plot_preds, score_type)
}
\arguments{
\item{plot_preds}{A string (vector).
The predictors to use when plotting HR.}

\item{score_type}{A string (vector). The score types in the analysis.
Available options include:
\itemize{
\item \code{CCI} for Charlson comorbidity index,
\item \code{EI} for Elixhauser index,
\item \code{PRS} for polygenic risk score,
\item \code{PheRS} for PheRS results score,
\item \code{MED} for medication data, and
\item \code{EDU} for educational level. The column for this needs to
be called \code{EDU} in the \code{pheno_data}.
\item \code{ZIP} for zip scores,
}}
}
\value{
A string vector of predictors to use when plotting hazard ratios.
}
\description{
This function returns the predictors to use when plotting hazard ratios.
If \code{plot_preds} is not NULL, it is returned. Otherwise, \code{score_type} is returned.
}
\author{
Kira E. Detrois
}
