% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{get_pred_string}
\alias{get_pred_string}
\title{Creates the predictor string for the Cox-PH model}
\usage{
get_pred_string(covs, score_type, pred_score = "SCORE_GROUP")
}
\arguments{
\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}

\item{pred_score}{A character. The score type to predict. For direct
regression \code{SCORE}, and for regression on the
risk group \code{SCORE_GROUP}.}
}
\value{
A character. The predictor string for the Cox-PH model.
}
\description{
Creates the predictor string for the Cox-PH model
}
\author{
Kira E. Detrois
}
