% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{get_pred_string}
\alias{get_pred_string}
\title{Creates the predictor string for the Cox-PH model}
\usage{
get_pred_string(covs, pred_score = "SCORE_GROUP")
}
\arguments{
\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}
}
\value{
A character. The predictor string for the Cox-PH model.

A character. The predictor string for the Cox-PH model.
}
\description{
Creates the predictor string for the Cox-PH model
}
\author{
Kira E. Detrois
Creates the predictor string for the Cox-PH model

Kira E. Detrois
}
