% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score_data.R
\name{get_study_med_data}
\alias{get_study_med_data}
\title{Calcualtes the MI score on the exposure window}
\usage{
get_study_med_data(elig_indv, atc_data)
}
\arguments{
\item{elig_indv}{A data.frame. The individuals which were eligble under
the current study setup. Needs to at least
contain the column defined in \code{ID}, \code{DATE_OF_BIRTH},
and \code{EXP_END_PERIOD}.}

\item{icd_data}{A data.frame with at least columns \code{ID}, \code{Event_age},
and \code{PRIMARY_ICD}.}
}
\value{
A data.frame with columns \code{ID}, and \code{SCORE}.
}
\description{
Calcualtes the MI score on the exposure window
}
\author{
Kira E. Detrois
}
