% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_descr.R
\name{get_study_subtitle}
\alias{get_study_subtitle}
\title{Creates a subtitle for a plot based on the current study setup}
\usage{
get_study_subtitle(study_setup)
}
\arguments{
\item{study_setup}{An S4 class representing the study setup.}
}
\value{
A character. The subtitle describing the study setup.
}
\description{
Creates a subtitle for a plot based on the current study setup
}
\author{
Kira E. Detrois
}
