% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_file_names.R
\name{get_surv_file_name}
\alias{get_surv_file_name}
\title{Creats the file name for endpoint specific score distribution plot}
\usage{
get_surv_file_name(surv_ana)
}
\value{
A character. The file name.
}
\description{
Creats the file name for endpoint specific score distribution plot
}
\author{
Kira E. Detrois
}
