% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_study_cidx.R
\name{get_surv_obj}
\alias{get_surv_obj}
\title{Creates a survival object for the selected endpoint}
\usage{
get_surv_obj(study_data, endpt)
}
\arguments{
\item{study_data}{A data.frame. The data for the Cox-PH model.}

\item{endpt}{A character. The current endpoint.}
}
\description{
Creates a survival object for the selected endpoint. With time
being the age in years until onset or the age at the end of the
observation period.
}
\author{
Kira E. Detrois
}
