% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_two_level_groups.R
\name{get_two_level_groups}
\alias{get_two_level_groups}
\title{Gets the risk score groups for each individual based on a single
cutoff}
\usage{
get_two_level_groups(score_data, cutoff)
}
\arguments{
\item{score_data}{A data.frame with at least columns \code{SCORE}, or
i.e. \code{J10_ASTHMA_PRS} for endpoint \code{J10_ASTHMA}
and score type \code{PRS}.}

\item{cutoff}{A numeric. The score cutoff value.}
}
\value{
A character. The group names.
}
\description{
Gets the risk score groups for each individual based on a single
cutoff
}
\author{
Kira E. Detrois
}
