% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot_adds.R
\name{ggplot_add_color_and_theme}
\alias{ggplot_add_color_and_theme}
\title{Add color and theme settings to the ggplot}
\usage{
ggplot_add_color_and_theme(
  plt,
  legend_name = "Predictor",
  color_labels,
  theme_base_size = 21
)
}
\arguments{
\item{plt}{A ggplot object.}

\item{legend_name}{A character. The legend name. Default: "Predictor".}

\item{color_labels}{A character (vector). The labels for the legend.}

\item{theme_base_size}{A numeric. The base font size for theme settings}
}
\value{
A ggplot object with added color and theme settings
}
\description{
Adds custom color settings and theme settings to a ggplot object
}
\author{
Kira E. Detrois
}
