% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_dfs}
\alias{join_dfs}
\title{Joins the two data.frames}
\usage{
join_dfs(endpt_data, score_data)
}
\value{
The joined data.frame
}
\description{
Joins the two data.frames and removes any individuals without
a score available.
}
\author{
Kira E. Detrois
}
