% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dfs.R
\name{join_dfs}
\alias{join_dfs}
\title{Joins phenotype and score data}
\usage{
join_dfs(pheno_data, score_data, score_type = "CCI")
}
\arguments{
\item{pheno_data}{A data.frame. The phenotype data. Needs at least column \code{ID}.}

\item{score_data}{A data.frame. The score data. Needs at least column \code{ID} and
if \code{score_type} contains \code{CCI} or \code{EI} then the selected columns.}

\item{score_type}{A string (vector). The score types used in the analysis.}
}
\value{
A tibble containing the merged phenotype and score data.
}
\description{
Joins the data.frames \code{pheno_data}, and \code{score_data} using
the \code{ID} variable.
}
\details{
Sets missing values for the \code{CCI}, or \code{EI} to 0.
}
\author{
Kira E. Detrois
}
