% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_dfs}
\alias{join_dfs}
\title{Joins the two data.frames}
\usage{
join_dfs(pheno_data, score_data, score_type = "CCI", endpt = NA_character_)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpts}.}

\item{score_data}{A data.frame. The score results for each individual.
Should have at least column defined in
\code{score_col_name} and column \code{ID}.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}
}
\value{
The joined data.frame
}
\description{
Joins the two data.frames and removes any individuals without
a score available.
}
\author{
Kira E. Detrois
}
