% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dfs.R
\name{join_dfs}
\alias{join_dfs}
\title{Joins the two data.frames}
\usage{
join_dfs(pheno_data, score_data, score_type = "CCI", endpt = NULL)
}
\arguments{
\item{pheno_data}{A data.frame. Left of the join. Needs at least column
\code{SCORE}.}

\item{score_data}{A data.frame. Left of the join. Needs at least column
\code{SCORE}.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}
}
\value{
The joined data.frame
}
\description{
Joins the two data.frames and removes any individuals without
a score available.
}
\author{
Kira E. Detrois
}
