% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dfs.R
\name{join_dfs}
\alias{join_dfs}
\title{Joins the two data.frames}
\usage{
join_dfs(study_data, score_data, score_type = "CCI", endpt = NULL)
}
\arguments{
\item{score_data}{A data.frame with at least columns \code{SCORE}, or
i.e. \code{J10_ASTHMA_PRS} for endpoint \code{J10_ASTHMA}
and score type \code{PRS}.}

\item{score_type}{A character. The score types used in the Cox-PH model.}

\item{endpt}{A character. The column name of the current endpoint
of interest.}
}
\value{
The joined data.frame
}
\description{
Joins the two data.frames and removes any individuals without
a score available.
}
\author{
Kira E. Detrois
}
