% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{log_msg_table}
\alias{log_msg_table}
\title{Creates a string of the current study setup}
\usage{
log_msg_table(score_group_tbl)
}
\arguments{
\item{score_group_tbl}{A named character.
The score cutoffs that should be written to the log.}
}
\value{
A tibble with columns \code{GROUP}, \code{SCORE_CUT}
}
\description{
Creates a string of the current study setup
}
\author{
Kira E. Detrois
}
