% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HRs.R
\name{plot_age_sd_hrs}
\alias{plot_age_sd_hrs}
\title{Plots the HR from the Cox-PH model of a forward study

Creates separate plots for each endpoint for the different age studies.
Plots both the risk group stratified and continuous HRs.}
\usage{
plot_age_sd_hrs(coxph_hrs, ana_details)
}
\arguments{
\item{coxph_hrs}{A tibble. The Cox-PH HR results. Needs
to at least contain the columns \code{ENDPOINT},
\code{HR}, \code{CI_NEG}, \code{CI_POS}, and \code{GROUP}.}
}
\description{
Plots the HR from the Cox-PH model of a forward study

Creates separate plots for each endpoint for the different age studies.
Plots both the risk group stratified and continuous HRs.
}
\author{
Kira E. Detrois
}
