% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_endpt_score_distr}
\alias{plot_endpt_score_distr}
\title{Plots a jittered scatter}
\usage{
plot_endpt_score_distr(pheno_score_data, score_type, study, write_res, res_dir)
}
\value{
A ggplot object
}
\description{
Plots a jittered scatter plot of the cases and controls scores
eligible under the current study setup.
}
\author{
Kira E. Detrois
}
