% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_endpt_score_distr}
\alias{plot_endpt_score_distr}
\title{Plots a jittered scatter}
\usage{
plot_endpt_score_distr(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{write_res},
\code{pheno_data}, \code{endpt}, \code{score_type}, and \code{exp_age},
and \code{exp_len}, \code{wash_len}, \code{obs_len}, and \code{res_dir}.}
}
\value{
A ggplot object
}
\description{
Plots a jittered scatter plot of the cases and controls scores
eligible under the current study setup.
}
\author{
Kira E. Detrois
}
