% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_distr.R
\name{plot_endpt_score_distr}
\alias{plot_endpt_score_distr}
\title{Plots a jittered scatter}
\usage{
plot_endpt_score_distr(
  score_data,
  score_type,
  study,
  obs_end = NULL,
  min_indvs = 5,
  write_res,
  res_dir
)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{study}{An S4 class representing the study setup.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\value{
A ggplot object
}
\description{
Plots a jittered scatter plot of the cases and controls scores
eligible under the current study setup.
}
\author{
Kira E. Detrois
}
