% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_score_distr.R
\name{plot_endpt_score_distr}
\alias{plot_endpt_score_distr}
\title{Plots a jittered scatter}
\usage{
plot_endpt_score_distr(score_data, surv_ana)
}
\arguments{
\item{score_data}{A data.frame with at least column \code{SCORE}.}

\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\value{
A ggplot object
}
\description{
Plots a jittered scatter plot of the cases and controls scores
eligible under the current study setup.
}
\author{
Kira E. Detrois
}
