% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HRs.R
\name{plot_endpt_sd_hr}
\alias{plot_endpt_sd_hr}
\title{Plots the HR from the Cox-PH model of a backwards study.

Creates a single plot with the different endpoints}
\usage{
plot_endpt_sd_hr(coxph_hrs, ana_details, sort_hrs, fig_height, fig_width)
}
\arguments{
\item{coxph_hrs}{A tibble. The Cox-PH HR results. Needs
to at least contain the columns \code{ENDPOINT},
\code{HR}, \code{CI_NEG}, \code{CI_POS}, and \code{GROUP}.}
}
\description{
Plots the HR from the Cox-PH model of a backwards study.

Creates a single plot with the different endpoints
}
\author{
Kira E. Detrois
}
