% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_score_distr}
\alias{plot_score_distr}
\title{Creates a plot of the score distribution}
\usage{
plot_score_distr(score_data, score_type, study, write_res, res_dir)
}
\value{
The ggplot object.
}
\description{
Creates a plot of the score distribution
}
\author{
Kira E. Detrois
}
