% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_score_distr}
\alias{plot_score_distr}
\title{Creates a plot of the score distribution}
\usage{
plot_score_distr(envir)
}
\arguments{
\item{envir}{A list with at least entries \code{write_res},
\code{score_data}, \code{score_type}, \code{exp_age}, and \code{exp_len},
and \code{res_dir}.}
}
\value{
The ggplot object.
}
\description{
Creates a plot of the score distribution
}
\author{
Kira E. Detrois
}
