% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_score_distr}
\alias{plot_score_distr}
\title{Creates a plot of the score distribution}
\usage{
plot_score_distr(score_data, score_type, study, write_res, res_dir)
}
\arguments{
\item{score_data}{A data.frame. Needs at least column \code{SCORE} if
\code{score_type = CCI}, or i.e. \code{J10_ASTHMA_PRS} for
\code{score_type = PRS} depending on the chosen \code{endpts}.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{study}{An S4 class representing the study setup.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\value{
The ggplot object.
}
\description{
Creates a plot of the score distribution
}
\author{
Kira E. Detrois
}
