% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess_score_data}
\alias{preprocess_score_data}
\title{Preprocesses the score data for the analysis}
\usage{
preprocess_score_data(score_data, score_col_name)
}
\arguments{
\item{score_data}{A data.frame with the score results for each individuals.
Should have at least column defined in \code{score_col_name}.}

\item{score_col_name}{A character. The name of the column with the scores.}
}
\value{
The data.frame with the score column renamed to \code{SCORE}.
}
\description{
Checks that \code{score_col_name} is contained in the data and renames the column
to \code{SCORE}.
}
\author{
Kira E. Detrois
}
