% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_hrs.R
\name{run_and_add_coxph_ana}
\alias{run_and_add_coxph_ana}
\title{Runs the cox-PH analysis and adds the results to the data.frame
\code{coxph_res}}
\usage{
run_and_add_coxph_ana(envir, predictor)
}
\arguments{
\item{envir}{A list with at least entries \code{coxph_res},
\code{score_type}, \code{endpt}, \code{pheno_score_data}.}

\item{predictor}{A character. The predictor for the survival
analyis string.}
}
\value{
A tibble the results data.fram \code{coxph_res} with added
columns from the analysis run.
}
\description{
Runs the cox-PH analysis and adds the results to the data.frame
\code{coxph_res}
}
