% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_ana.R
\name{run_coxph_ana}
\alias{run_coxph_ana}
\title{Fits the cox-ph model for the current endpoint}
\usage{
run_coxph_ana(data, endpt, predictor = "SCORE_GROUP")
}
\arguments{
\item{data}{TODO}

\item{endpt}{A string. The column name of the current endpoint of
interest.}
}
\description{
Fits the cox-ph model for the current endpoint
}
