% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_ana.R
\name{run_coxph_ana}
\alias{run_coxph_ana}
\title{Fits the Cox model for a given endpoint and predictor}
\usage{
run_coxph_ana(pheno_score_data, endpt, predictor = "SCORE_GROUP")
}
\arguments{
\item{pheno_score_data}{A data.frame TODO.}

\item{endpt}{A string. The column name of the current endpoint of
interest.}
}
\description{
Fits the Cox model for a given endpoint and predictor
}
