% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coxph_ana.R
\name{run_coxph_ana}
\alias{run_coxph_ana}
\title{Fits the Cox model for a given endpoint and predictor}
\usage{
run_coxph_ana(
  pheno_score_data,
  endpt,
  predictor = "SCORE_GROUP",
  write_res = FALSE
)
}
\arguments{
\item{pheno_score_data}{A data.frame with at least columns \code{DATE_OF_BIRTH}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_AGE_DAYS}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}. And the column defined in \code{predictor}.}

\item{endpt}{A character. The current enpoint of interest.}

\item{predictor}{A character. The predictor for the survival analysis.}
}
\description{
Fits the Cox model for a given endpoint and predictor
}
\author{
Kira E. Detrois
}
