% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_surv_studies.R
\name{run_surv_studies}
\alias{run_surv_studies}
\title{Fits a Cox-PH model for each endpoint}
\usage{
run_surv_studies(
  pheno_data,
  endpts_indvs_mat = NULL,
  icd_data = NULL,
  atc_data = NULL,
  prs_data = NULL,
  phers_data = NULL,
  score_type,
  create_score_combos = FALSE,
  plot_preds = NULL,
  study_type = "backward",
  endpts = c("J10_ASTHMA"),
  exp_ages = 0,
  exp_len = NULL,
  wash_len = 2,
  obs_len = 8,
  obs_end_date = as.Date("2021/01/01"),
  down_fctr = NA_integer_,
  ancs = NA_character_,
  obs_age_range = c(32, 80),
  covs = c("SEX", "YEAR_OF_BIRTH"),
  min_indvs = 5,
  res_descr = "",
  write_res = FALSE,
  res_dir = NULL
)
}
\arguments{
\item{pheno_data}{A data.frame with the phenotypic data on the individuals.
Should at least contain the columns:
\code{ID}, the columns specified in \code{covs} and
i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_AGE_FROM_BASE} where
the columns are the study endpoint and the age
from baseline, which will differ depending on the chosen
\code{endpts}.}

\item{endpts_indvs_mat}{A data.frame. A matrix with the individuals that can
be used for each endpoint.}

\item{icd_data}{A data.frame. A data.frame with the ICD codes of the
individuals.}

\item{atc_data}{A data.frame. A data.frame with the ATC codes of the
individuals.}

\item{prs_data}{A data.frame. A data.frame with the PRS scores of the
individuals.}

\item{phers_data}{A data.frame. A data.frame with the PheRS scores of the
individuals.}

\item{score_type}{A character (vector).
The score types used in the Cox-PH model.}

\item{plot_preds}{A character (vector).
The predictors to use when plotting HR.}

\item{study_type}{A character. Can be either \code{forward} or \code{backward}.
\code{forward} considers individuals of a certain age
and a set exposure, washout, and observation periods
calcualted onwards from this age. It can simply be created
by creating a S4 study object.
\code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point. The exposure period will
be different for each individual depending on their birth
date. This setup can be created, using the function
\code{\link{get_backward_study}}.}

\item{endpts}{A character (vector). The column names of the endpoints of
interest.}

\item{exp_ages}{An integer (vector). Ages at which exposure period
starts (in years).}

\item{exp_len}{An integer. Length of the exposure period ^
(in years).}

\item{wash_len}{An integer. Length of the washout period
(in years).}

\item{obs_len}{An integer. Length of the prediction period
(in years).}

\item{obs_end_date}{A Date. The end of the observation period. Needed
for \code{backward} studies. If not set the end will
be the most recent date across the diagnosis columns.}

\item{down_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{ancs}{A character (vector). The ancestries to consider.}

\item{obs_age_range}{A numeric. The age range of individuals in the observation
period. Inclusive interval.}

\item{covs}{A vector of characters. The column names of the covariates
to add to the predictor of the Cox-PH model.}

\item{min_indvs}{An integer. The minimum number of individuals each
group in the analyses needs to have. This is important
for being able to export data from the FinnGen Sanbdox.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}

\item{score_type_combos}{A boolean, whether or not to create all possible
score type combinations from the score_type vector.
see function \link[IUtils:get_all_possible_score_type_combs]{IUtils::get_all_possible_score_type_combs}.}
}
\description{
This function runs survival studies for a given endpoint.It fits a
Cox proportional hazards model to the data and
calculates the C-index. The resulting HRs and c-indices are written
to files if wanted.
}
