% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_coxph.R
\name{scale_preds}
\alias{scale_preds}
\title{Scales the given variables in the score data}
\usage{
scale_preds(preds, study_data)
}
\arguments{
\item{preds}{A character vector of the variables to be scaled}

\item{study_data}{A data frame with the variables to be scaled}
}
\value{
A data frame with the scaled variables
}
\description{
Scales the given variables in the score data
}
