% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ana.R
\name{setPreds}
\alias{setPreds}
\title{Sets the presictors of the S4 surv_ana object to a new value}
\usage{
setPreds(.Object, crnt_score_types)
}
\arguments{
\item{.Object}{The S4 \link{surv_ana} object.}

\item{crnt_score_types}{A string (vector). The current score types.}
}
\value{
The updated S4 object
}
\description{
Sets the presictors of the S4 surv_ana object to a new value
}
\author{
Kira E. Detrois
}
