% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ana.R
\docType{class}
\name{surv_ana-class}
\alias{surv_ana-class}
\alias{surv_ana}
\title{An S4 class representing the survival analysis}
\description{
An S4 class representing the survival analysis
}
\section{Slots}{

\describe{
\item{\code{study}}{A S4 study object. The current study setups for on whicht
the eligible indiviuals are selected.}

\item{\code{study_data}}{A dataframe with the phenotypic and score data
of all eligible individuals under
the current study setup and their scores.}

\item{\code{min_indvs}}{An integer. The minimum number of individuals each
group in the analyses needs to have.}

\item{\code{preds}}{A character (vector). The names of the covariates
to use in the Cox-PH model.}

\item{\code{plot_preds}}{A character (vector).
The predictors to use when plotting the HRs.}

\item{\code{res_descr}}{A character. An additional descriptive that can be
used for distinguishing the results files.}

\item{\code{write_res}}{A boolean. Defines whether to save the results to
files. Default: FALSE.}

\item{\code{res_dir}}{A character. The directory to write the results and
logs to. Default: NA.}
}}

\author{
Kira E. Detrois
}
