% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_logs.R
\name{write_pheno_score_files}
\alias{write_pheno_score_files}
\title{Writes results and log file}
\usage{
write_pheno_score_files(pheno_score_data, study_setup, endpt, surv_ana)
}
\arguments{
\item{surv_ana}{An S4 study object. The current survival analysis.}
}
\description{
Prints to console if something goes wrong.
}
\author{
Kira E. Detrois
}
