% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_res_file}
\alias{write_res_file}
\title{Writes model results to a tab-delim file}
\usage{
write_res_file(coxph_res_tib, score_type, study, write_res, res_dir)
}
\arguments{
\item{coxph_res_tib}{A tibble with the results for the current endpoints.}

\item{score_type}{A character. The name of the score used for the model,
i.e. CCI, or PheRS.}

\item{study}{An S4 class representing the study setup.}

\item{write_res}{A boolean. Defines whether to save the results to
files.}

\item{res_dir}{A character. The directory to write the results and
log to.}
}
\description{
Writes model results to a tab-delim file
}
\author{
Kira E. Detrois
}
