% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_files.R
\name{write_res_files}
\alias{write_res_files}
\title{Writes model results to a tab-delim file}
\usage{
write_res_files(
  hr_res,
  c_idxs_res,
  study_setup,
  surv_ana,
  score_types,
  is_first_endpt
)
}
\arguments{
\item{hr_res}{A tibble containing the hazard ratios for each exposure age.}

\item{c_idxs_res}{A data.frame. The c-index results.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link[Istudy:study_setup]{Istudy::study_setup}.}

\item{surv_ana}{An S4 \code{surv_ana} object. The current survival analysis setup.
See class definition \link{surv_ana}.}
}
\description{
Writes model results to a tab-delim file
}
\author{
Kira E. Detrois
}
