% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_files.R
\name{write_res_files}
\alias{write_res_files}
\title{Writes model results to a tab-delim file}
\usage{
write_res_files(endpt_hrs_tib, endpt_c_idxs_tib, surv_ana)
}
\arguments{
\item{endpt_hrs_tib}{A tibble with the results for previous endpts.}

\item{endpt_c_idxs_tib}{A tibble with the results for previous endpts.}
}
\description{
Writes model results to a tab-delim file
}
\author{
Kira E. Detrois
}
