% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_write_files.R
\name{write_score_groups_to_log}
\alias{write_score_groups_to_log}
\title{Write score cut-off values and groups to log}
\usage{
write_score_groups_to_log(score_group_tbl, surv_ana)
}
\arguments{
\item{score_group_tbl}{A named numeric. The risk score cuts for the
different quantiles.}

\item{surv_ana}{An S4 class representing the current survival analysis
setup.}
}
\description{
Write score cut-off values and groups to log
}
\author{
Kira E. Detrois
}
